Split storage tests by subsystem
