+++ crates/truthdb-core/src/storage/mod.rs
@@ -8500,3884 +8500,4 @@ fn subtract_run(runs: Vec<(u64, u64)>, start: u64, len: u64) -> Vec<(u64, u64)> 
 }
 
 #[cfg(test)]
-mod tests {
-    use std::io::{Read, Seek, SeekFrom, Write};
-    use std::path::{Path, PathBuf};
-
-    use super::*;
-    use crate::storage_layout::{
-        FILE_HEADER_SIZE, FILE_VERSION, FILE_VERSION_V1, SUPERBLOCK_SIZE, WAL_ENTRY_TYPE_RECORD,
-        WalEntryFooter, WalEntryHeader, wal_entry_padded_len, wal_payload_crc,
-    };
-
-    /// Small ring so wrap/full paths are cheap to reach.
-    const TEST_WAL_BYTES: u64 = 64 * 1024;
-
-    fn test_storage_options() -> StorageOptions {
-        StorageOptions {
-            size_gib: 1,
-            wal_ratio: 0.05,
-            metadata_ratio: 0.08,
-            snapshot_ratio: 0.02,
-            allocator_ratio: 0.02,
-            reserved_ratio: 0.17,
-            default_collation: None,
-        }
-    }
-
-    #[test]
-    fn backup_is_single_flight_and_releases_its_hold() {
-        let path = unique_temp_path("backup-hold");
-        let bak = unique_temp_path("backup-hold-bak");
-        let bak2 = unique_temp_path("backup-hold-bak2");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-
-        // A second backup while a hold is set is rejected, and — critically —
-        // must NOT clear the in-flight backup's hold (the guard arms only after
-        // begin_backup succeeds, so a rejected backup touches nothing).
-        storage.lock().register_backup_hold(42);
-        assert!(matches!(
-            storage.backup_full(&bak),
-            Err(StorageError::BackupInProgress)
-        ));
-        assert_eq!(
-            storage.lock().truncation_gate.backup,
-            Some(42),
-            "a rejected backup leaves the existing hold intact"
-        );
-        storage.lock().release_backup_hold();
-
-        // A successful backup releases its own hold, so a second one succeeds.
-        storage.backup_full(&bak).expect("first backup");
-        assert_eq!(
-            storage.lock().truncation_gate.backup,
-            None,
-            "the hold is released after a successful backup"
-        );
-        storage
-            .backup_full(&bak2)
-            .expect("second backup after release");
-
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-        let _ = std::fs::remove_file(bak);
-        let _ = std::fs::remove_file(bak2);
-    }
-
-    #[test]
-    fn log_backup_orphans_when_recovery_flips_to_simple_mid_flight() {
-        // Reproduces the race the lock-dance opened: BACKUP LOG releases the
-        // storage lock to write its archive; a concurrent ALTER ... SET RECOVERY
-        // SIMPLE releases the log hold and a checkpoint advances the head; phase
-        // 3 must then ORPHAN the backup rather than re-arm the hold below head.
-        let path = unique_temp_path("backuplog-orphan");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        storage
-            .rel_set_db_options(None, None, Some(true))
-            .expect("enable FULL");
-        for seq in 0..8 {
-            storage
-                .append_wal_entry(WAL_ENTRY_TYPE_RECORD, 1, seq, b"logbytes")
-                .expect("append");
-        }
-        // Phase 1: capture the range under the lock (marker = start).
-        let (_, start, end, _log) = storage.lock().begin_log_backup(true, false).expect("begin");
-        assert!(end > start);
-        // The concurrent ALTER SET RECOVERY SIMPLE during the unlocked window.
-        storage
-            .rel_set_db_options(None, None, Some(false))
-            .expect("disable FULL");
-        assert_eq!(storage.log_backup_hold(), None, "SIMPLE released the hold");
-        // A checkpoint now advances the head to the tail (no hold pins it).
-        storage
-            .write_checkpoint(b"cp", 1, 2, 1)
-            .expect("checkpoint after SIMPLE");
-        assert!(
-            storage.wal_head() > start,
-            "the head advanced past the old marker"
-        );
-        // Phase 3: finish must orphan (recovery_full is false) — no re-arm.
-        storage
-            .lock()
-            .finish_log_backup(start, end)
-            .expect("finish orphans cleanly");
-        assert_eq!(
-            storage.log_backup_hold(),
-            None,
-            "finish did not re-arm the hold on the now-SIMPLE database"
-        );
-        // The next checkpoint must not move the head backward (would panic
-        // without the orphan guard).
-        storage
-            .write_checkpoint(b"cp2", 2, 3, 2)
-            .expect("checkpoint after orphan does not panic");
-        storage.lock().cancel_log_backup();
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-    }
-
-    #[test]
-    fn a_page_freed_since_the_backup_began_is_not_treated_as_corrupt() {
-        let path = unique_temp_path("backup-freed");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        {
-            let mut file = storage.lock();
-            let page = file.allocator.allocate(1).expect("allocate a page");
-            assert!(
-                file.page_is_live_regular(page).unwrap(),
-                "a live, allocated data page is regular (a checksum failure there is real corruption)"
-            );
-            file.allocator.free(page, 1);
-            assert!(
-                !file.page_is_live_regular(page).unwrap(),
-                "a page freed since the backup began is tolerated, not flagged corrupt"
-            );
-        }
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-    }
-
-    /// ALTER TABLE ADD survives a restart — the widened schema and the frozen
-    /// fills are one durable statement — and a failure mid-rewrite rolls the
-    /// whole ALTER back: old schema, old rows, fully readable.
-    #[test]
-    fn alter_add_column_is_durable_and_atomic() {
-        use crate::engine::{StatementResult, TxnContext, execute_batch};
-
-        let path = unique_temp_path("alter-add");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        for sql in [
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, name NVARCHAR(20))",
-            "INSERT INTO t VALUES (1, 'one'), (2, 'two')",
-            "ALTER TABLE t ADD score INT DEFAULT 7",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        drop(storage);
-
-        // Reopen: the widened schema and the frozen fill survived.
-        let storage = Storage::open(path.clone()).expect("reopen");
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(&storage, "SELECT id, score FROM t ORDER BY id", &mut ctx);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => assert_eq!(
-                rowset.rows,
-                vec![
-                    vec![Datum::Int(1), Datum::Int(7)],
-                    vec![Datum::Int(2), Datum::Int(7)],
-                ]
-            ),
-            other => panic!("expected rows, got {other:?}"),
-        }
-
-        // A failure mid-rewrite (fault injection inside the ALTER's statement)
-        // rolls the whole thing back: the new column does not exist and the
-        // old rows are intact.
-        crate::relstore::ctx::FAIL_APPLY_OPS_AFTER.with(|c| c.set(Some(1)));
-        let outcome = execute_batch(&storage, "ALTER TABLE t ADD flag BIT DEFAULT 1", &mut ctx);
-        crate::relstore::ctx::FAIL_APPLY_OPS_AFTER.with(|c| c.set(None));
-        assert!(outcome.error.is_some(), "the injected failure must surface");
-
-        let outcome = execute_batch(&storage, "SELECT flag FROM t", &mut ctx);
-        assert!(
-            outcome.error.is_some(),
-            "the rolled-back column must not exist: {:?}",
-            outcome.results
-        );
-        let outcome = execute_batch(&storage, "SELECT id, score FROM t ORDER BY id", &mut ctx);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => assert_eq!(rowset.rows.len(), 2),
-            other => panic!("expected rows, got {other:?}"),
-        }
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// The EXEC text path's two load-bearing lock properties, pinned: a
-    /// variable @stmt cannot be analyzed up front and takes the conservative
-    /// database-exclusive lock, and isolation escalation crosses the EXEC
-    /// boundary in both directions (mutating either previously went green).
-    #[test]
-    fn exec_lock_analysis_never_under_locks() {
-        use crate::engine::{Isolation, TxnContext, execute_batch};
-        use crate::lock::{LockMode, Resource};
-
-        let path = unique_temp_path("exec-locks");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(
-            &storage,
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-
-        // A variable statement text is unknowable before it runs: the batch
-        // locks the database exclusively rather than ever under-locking.
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "DECLARE @s NVARCHAR(50) = N'SELECT v FROM t'; EXEC sp_executesql @s",
-            Isolation::ReadCommitted,
-        );
-        assert!(
-            needs.contains(&(Resource::Database, LockMode::Exclusive)),
-            "variable @stmt must take Database X: {needs:?}"
-        );
-
-        // Direction 1: a SET raise BEFORE the EXEC locks the inner reads.
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "SET TRANSACTION ISOLATION LEVEL SERIALIZABLE; EXEC sp_executesql N'SELECT v FROM t'",
-            Isolation::ReadUncommitted,
-        );
-        assert!(
-            needs
-                .iter()
-                .any(|(r, m)| matches!(r, Resource::Table(_)) && *m == LockMode::Shared),
-            "escalated isolation must lock the inner SELECT: {needs:?}"
-        );
-
-        // Direction 2 (intra-EXEC): a SET raise INSIDE the literal locks the
-        // statements after it inside the same literal.
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "EXEC sp_executesql N'SET TRANSACTION ISOLATION LEVEL SERIALIZABLE; SELECT v FROM t'",
-            Isolation::ReadUncommitted,
-        );
-        assert!(
-            needs
-                .iter()
-                .any(|(r, m)| matches!(r, Resource::Table(_)) && *m == LockMode::Shared),
-            "an inner SET raise must lock the inner reads: {needs:?}"
-        );
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// The Stage 13 database options round-trip the superblock: they survive
-    /// a reopen, and a checkpoint — which rebuilds both superblocks from
-    /// scratch — must carry them forward rather than silently resetting them.
-    #[test]
-    fn db_options_persist_across_reopen_and_checkpoint() {
-        use crate::engine::{TxnContext, execute_batch};
-
-        let path = unique_temp_path("db-options");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        assert!(!storage.rcsi_enabled());
-        assert!(!storage.snapshot_isolation_allowed());
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(
-            &storage,
-            "ALTER DATABASE CURRENT SET READ_COMMITTED_SNAPSHOT ON, ALLOW_SNAPSHOT_ISOLATION ON",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        assert!(storage.rcsi_enabled());
-        assert!(storage.snapshot_isolation_allowed());
-        drop(storage);
-
-        let storage = Storage::open(path.clone()).expect("reopen");
-        assert!(storage.rcsi_enabled(), "RCSI survives a restart");
-        assert!(storage.snapshot_isolation_allowed());
-
-        // One option off, then a checkpoint, then a reopen: the checkpoint's
-        // fresh superblocks must keep the surviving option.
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(
-            &storage,
-            "ALTER DATABASE CURRENT SET ALLOW_SNAPSHOT_ISOLATION OFF",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        storage
-            .write_checkpoint(b"cp", 1, 2, 1)
-            .expect("checkpoint");
-        drop(storage);
-        let storage = Storage::open(path.clone()).expect("reopen after checkpoint");
-        assert!(
-            storage.rcsi_enabled(),
-            "the checkpoint must not reset the option"
-        );
-        assert!(!storage.snapshot_isolation_allowed());
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// A READ COMMITTED SELECT under RCSI takes only Database IS — no Table
-    /// S — which is the entire readers-don't-block mechanism; and the other
-    /// levels are untouched by the option.
-    /// Lock analysis descends control-flow bodies: a WHILE body's INSERT and
-    /// an IF condition's EXISTS table are in the batch's up-front lock set.
-    /// EXEC of a user procedure locks the STORED BODY's tables up front —
-    /// parsed with the in-procedure grammar (a plain parse would 178 on
-    /// `RETURN <value>`, yield no locks, and the body would run unlocked).
-    /// Recursive procedures terminate analysis via the visited set.
-    #[test]
-    fn analyze_locks_covers_procedure_bodies() {
-        use crate::engine::{Isolation, TxnContext, execute_batch};
-        use crate::lock::{LockMode, Resource};
-
-        let path = unique_temp_path("proc-locks");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        for sql in [
-            "CREATE TABLE plt (id INT NOT NULL PRIMARY KEY)",
-            "CREATE PROCEDURE writer @v INT AS INSERT INTO plt VALUES (@v); \
-             EXEC writer @v; RETURN 5",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "EXEC writer 1",
-            Isolation::ReadCommitted,
-        );
-        assert!(
-            needs.iter().any(
-                |(r, m)| matches!(r, Resource::Table(_)) && *m == LockMode::Exclusive
-                    || matches!(r, Resource::Row(..))
-            ),
-            "the recursive body's INSERT locks its table (and analysis \
-             terminated): {needs:?}"
-        );
-        // An unknown procedure contributes no locks (2812 at execution).
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "EXEC no_such_proc",
-            Isolation::ReadCommitted,
-        );
-        assert!(
-            !needs
-                .iter()
-                .any(|(r, _)| matches!(r, Resource::Table(_) | Resource::Row(..))),
-            "unknown proc: {needs:?}"
-        );
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Adversarial review probe: the visited set dedups a procedure's lock
-    /// contribution, but a body's lock set DEPENDS on the effective
-    /// isolation. Under RCSI, `EXEC pread` analyzed first contributes only
-    /// Database IS (versioned read); a later `EXEC pser` — whose body raises
-    /// to SERIALIZABLE and EXECs pread — finds pread already visited and
-    /// skips it, so the lock-based re-analysis (Table S) is dropped. At
-    /// execution the SET is live inside pser and pread's SELECT reads
-    /// lock-based with no Table S held: the 2PL under-lock class.
-    #[test]
-    fn review_poc_analyze_locks_procedure_reanalyzed_under_escalated_isolation() {
-        use crate::engine::{Isolation, TxnContext, execute_batch};
-        use crate::lock::{LockMode, Resource};
-
-        let path = unique_temp_path("proc-visited-isolation");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        for sql in [
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)",
-            "ALTER DATABASE CURRENT SET READ_COMMITTED_SNAPSHOT ON",
-            "CREATE PROCEDURE pread AS SELECT v FROM t",
-            "CREATE PROCEDURE pser AS \
-             SET TRANSACTION ISOLATION LEVEL SERIALIZABLE; EXEC pread",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        let table_s = |needs: &[(Resource, LockMode)]| {
-            needs
-                .iter()
-                .any(|(r, m)| matches!(r, Resource::Table(_)) && *m == LockMode::Shared)
-        };
-        // Control: analyzed alone, the escalated body read-locks.
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "EXEC pser",
-            Isolation::ReadCommitted,
-        );
-        assert!(
-            table_s(&needs),
-            "control: pser's escalated body takes Table S: {needs:?}"
-        );
-        // The seam: pread analyzed first under the versioned regime, then
-        // pser's escalated re-entry is dropped by the visited set.
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "EXEC pread; EXEC pser",
-            Isolation::ReadCommitted,
-        );
-        assert!(
-            table_s(&needs),
-            "pser still runs pread's SELECT under SERIALIZABLE — Table S \
-             must be in the up-front set: {needs:?}"
-        );
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Adversarial review probe: a stored procedure EXEC'd from INSIDE a
-    /// dynamic-SQL literal is still resolved by lock analysis (the literal
-    /// recursion's Exec arm hits the procedure branch).
-    #[test]
-    fn review_poc_analyze_locks_procedure_via_dynamic_sql() {
-        use crate::engine::{Isolation, TxnContext, execute_batch};
-        use crate::lock::{LockMode, Resource};
-
-        let path = unique_temp_path("proc-dyn-locks");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        for sql in [
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)",
-            "CREATE PROCEDURE wtr @v INT AS INSERT INTO t VALUES (@v)",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "EXEC sp_executesql N'EXEC wtr 1'",
-            Isolation::ReadCommitted,
-        );
-        assert!(
-            needs.iter().any(|(r, m)| matches!(r, Resource::Table(_))
-                && matches!(m, LockMode::IntentExclusive | LockMode::Exclusive)
-                || matches!(r, Resource::Row(..))),
-            "the proc body's INSERT is in the up-front set via the literal \
-             path: {needs:?}"
-        );
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Adversarial review probe: analysis resolves the catalog FIRST, but
-    /// execution checks the sp_executesql builtin FIRST. A user procedure
-    /// named sp_executesql makes the two disagree: analysis analyzes the
-    /// user body, execution runs the builtin over the literal — whose locks
-    /// were never analyzed (under-lock). Either the CREATE must be refused
-    /// or the analysis must mirror execution's builtin-first order.
-    #[test]
-    fn review_poc_user_procedure_named_sp_executesql_cannot_shadow_builtin() {
-        use crate::engine::{Isolation, TxnContext, execute_batch};
-        use crate::lock::{LockMode, Resource};
-
-        let path = unique_temp_path("proc-spexec-shadow");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(
-            &storage,
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        let outcome = execute_batch(
-            &storage,
-            "CREATE PROCEDURE sp_executesql AS SELECT 1 AS n",
-            &mut ctx,
-        );
-        if outcome.error.is_none() {
-            // The shadow exists: analysis and execution must still agree.
-            // Execution runs the BUILTIN (its name check comes first), so the
-            // literal's INSERT locks must be in the analyzed set.
-            let needs = crate::engine::analyze_locks(
-                &storage,
-                crate::relstore::catalog::DEFAULT_DATABASE_ID,
-                "EXEC sp_executesql N'INSERT INTO t VALUES (1)'",
-                Isolation::ReadCommitted,
-            );
-            assert!(
-                needs.iter().any(|(r, m)| matches!(r, Resource::Table(_))
-                    && matches!(m, LockMode::IntentExclusive | LockMode::Exclusive)
-                    || matches!(r, Resource::Row(..))
-                    || (matches!(r, Resource::Database) && *m == LockMode::Exclusive)),
-                "execution runs the builtin INSERT; analysis followed the \
-                 user proc's body instead: {needs:?}"
-            );
-        }
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    #[test]
-    fn analyze_locks_descends_control_flow() {
-        use crate::engine::{Isolation, TxnContext, execute_batch};
-        use crate::lock::{LockMode, Resource};
-
-        let path = unique_temp_path("flow-locks");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(
-            &storage,
-            "CREATE TABLE locked_t (id INT NOT NULL PRIMARY KEY)",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "DECLARE @i INT = 0; WHILE @i < 3 BEGIN INSERT INTO locked_t VALUES (@i); \
-             SET @i = @i + 1; END",
-            Isolation::ReadCommitted,
-        );
-        assert!(
-            needs.iter().any(
-                |(r, m)| matches!(r, Resource::Table(_)) && *m == LockMode::Exclusive
-                    || matches!(r, Resource::Row(..))
-            ),
-            "the WHILE body's INSERT locks its table: {needs:?}"
-        );
-
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "IF EXISTS (SELECT * FROM locked_t) SELECT 1",
-            Isolation::ReadCommitted,
-        );
-        assert!(
-            needs
-                .iter()
-                .any(|(r, m)| matches!(r, Resource::Table(_)) && *m == LockMode::Shared),
-            "the IF condition's EXISTS table takes Table S: {needs:?}"
-        );
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Adversarial probes (control-flow review): condition shapes whose table
-    /// reads must be in the up-front lock set — a WHILE condition, a derived
-    /// table and an IN-subquery inside the condition, a CASE-wrapped EXISTS,
-    /// a view, an untaken ELSE branch's write, and an EXEC literal inside a
-    /// WHILE body.
-    #[test]
-    fn cf_review_analyze_locks_condition_shapes() {
-        use crate::engine::{Isolation, TxnContext, execute_batch};
-        use crate::lock::{LockMode, Resource};
-
-        let path = unique_temp_path("cf-flow-lock-shapes");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        for sql in [
-            "CREATE TABLE lt (id INT NOT NULL PRIMARY KEY)",
-            "CREATE VIEW lv AS SELECT id FROM lt",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        let table_s = |needs: &[(Resource, LockMode)]| {
-            needs
-                .iter()
-                .any(|(r, m)| matches!(r, Resource::Table(_)) && *m == LockMode::Shared)
-        };
-        let table_write = |needs: &[(Resource, LockMode)]| {
-            needs.iter().any(|(r, m)| {
-                matches!(r, Resource::Table(_)) && *m == LockMode::Exclusive
-                    || matches!(r, Resource::Row(..))
-            })
-        };
-        for sql in [
-            "WHILE EXISTS (SELECT * FROM lt) SELECT 1",
-            "IF EXISTS (SELECT * FROM (SELECT id FROM lt) d) SELECT 1",
-            "IF 1 IN (SELECT id FROM lt) SELECT 1",
-            "IF CASE WHEN EXISTS (SELECT * FROM lt) THEN 1 ELSE 0 END = 1 SELECT 1",
-            "IF EXISTS (SELECT * FROM lv) SELECT 1",
-            "IF (SELECT COUNT(*) FROM lt) = 0 SELECT 1",
-        ] {
-            let needs = crate::engine::analyze_locks(
-                &storage,
-                crate::relstore::catalog::DEFAULT_DATABASE_ID,
-                sql,
-                Isolation::ReadCommitted,
-            );
-            assert!(
-                table_s(&needs),
-                "{sql}: condition read takes Table S: {needs:?}"
-            );
-        }
-        // Both IF branches analyze — an untaken ELSE's INSERT is still locked.
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "IF 1 = 2 SELECT 1 ELSE INSERT INTO lt VALUES (9)",
-            Isolation::ReadCommitted,
-        );
-        assert!(
-            table_write(&needs),
-            "the ELSE branch's INSERT locks its table: {needs:?}"
-        );
-        // An EXEC literal inside a WHILE body analyzes through the Exec arm.
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "DECLARE @i INT = 0; WHILE @i < 1 BEGIN \
-             EXEC sp_executesql N'INSERT INTO lt VALUES (7)'; SET @i = @i + 1; END",
-            Isolation::ReadCommitted,
-        );
-        assert!(
-            table_write(&needs),
-            "the EXEC'd INSERT inside the loop locks its table: {needs:?}"
-        );
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// A CTE inside an IF condition's subquery: the executor inlines it and
-    /// reads the base table (engine.rs pins that half), so analysis must lock
-    /// that table like the Select arm does — the expectation here is the
-    /// FIXED behavior.
-    #[test]
-    fn cf_review_analyze_locks_condition_cte() {
-        use crate::engine::{Isolation, TxnContext, execute_batch};
-        use crate::lock::{LockMode, Resource};
-
-        let path = unique_temp_path("cf-flow-lock-cte");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(
-            &storage,
-            "CREATE TABLE lt (id INT NOT NULL PRIMARY KEY)",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "IF EXISTS (WITH x AS (SELECT id FROM lt) SELECT id FROM x) SELECT 1",
-            Isolation::ReadCommitted,
-        );
-        assert!(
-            needs
-                .iter()
-                .any(|(r, m)| matches!(r, Resource::Table(_)) && *m == LockMode::Shared),
-            "the CTE's base table is read at runtime and must be locked: {needs:?}"
-        );
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    #[test]
-    fn analyze_locks_drops_table_s_under_rcsi() {
-        use crate::engine::{Isolation, TxnContext, execute_batch};
-        use crate::lock::{LockMode, Resource};
-
-        let path = unique_temp_path("rcsi-locks");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(
-            &storage,
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-
-        let table_s = |needs: &[(Resource, LockMode)]| {
-            needs
-                .iter()
-                .any(|(r, m)| matches!(r, Resource::Table(_)) && *m == LockMode::Shared)
-        };
-
-        // Off: the SELECT read-locks, as ever.
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "SELECT v FROM t",
-            Isolation::ReadCommitted,
-        );
-        assert!(
-            table_s(&needs),
-            "without RCSI a RC SELECT takes Table S: {needs:?}"
-        );
-
-        let outcome = execute_batch(
-            &storage,
-            "ALTER DATABASE CURRENT SET READ_COMMITTED_SNAPSHOT ON",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-
-        // On: Database IS only — the DDL fence — and no Table S.
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "SELECT v FROM t",
-            Isolation::ReadCommitted,
-        );
-        assert!(
-            !table_s(&needs),
-            "under RCSI a RC SELECT takes no Table S: {needs:?}"
-        );
-        assert!(
-            needs.contains(&(Resource::Database, LockMode::IntentShared)),
-            "the Database IS fence stays: {needs:?}"
-        );
-
-        // The other levels are untouched: RR still read-locks, RU still
-        // takes nothing, and a batch that raises isolation falls back to
-        // locking even though the session level is RC.
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "SELECT v FROM t",
-            Isolation::RepeatableRead,
-        );
-        assert!(table_s(&needs), "RR is not versioned: {needs:?}");
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "SELECT v FROM t",
-            Isolation::ReadUncommitted,
-        );
-        assert!(needs.is_empty(), "RU takes no locks at all: {needs:?}");
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "SET TRANSACTION ISOLATION LEVEL SERIALIZABLE; SELECT v FROM t",
-            Isolation::ReadCommitted,
-        );
-        assert!(
-            table_s(&needs),
-            "a raising SET disables the snapshot path: {needs:?}"
-        );
-
-        // SNAPSHOT isolation is versioned regardless of RCSI: Database IS
-        // only, and the EXEC-literal recursion preserves the level (the
-        // #120 review's collapse bug, from the other direction).
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "SELECT v FROM t",
-            Isolation::Snapshot,
-        );
-        assert!(
-            !table_s(&needs),
-            "SNAPSHOT reads take no Table S: {needs:?}"
-        );
-        assert!(needs.contains(&(Resource::Database, LockMode::IntentShared)));
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "EXEC sp_executesql N'SELECT v FROM t'",
-            Isolation::Snapshot,
-        );
-        assert!(
-            !table_s(&needs),
-            "the recursion must not turn SNAPSHOT into a locking level: {needs:?}"
-        );
-        // ...and a SET SNAPSHOT inside a batch is not a lock-escalating
-        // raise, but the batch still holds the Database IS fence.
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "SET TRANSACTION ISOLATION LEVEL SNAPSHOT; SELECT v FROM t",
-            Isolation::ReadUncommitted,
-        );
-        assert!(
-            needs.contains(&(Resource::Database, LockMode::IntentShared)),
-            "SET SNAPSHOT from RU keeps the DDL fence: {needs:?}"
-        );
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// A covering (INCLUDE) seek must serve the snapshot's image, not the
-    /// index leaf's freshly-updated include payload — pinned on the covering
-    /// path itself via the covering-scan counter.
-    #[test]
-    fn covering_seek_serves_the_snapshot_image() {
-        use crate::engine::{StatementResult, TxnContext, execute_batch};
-
-        let path = unique_temp_path("rcsi-covering");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut writer = TxnContext::default();
-        let mut fill = String::from("INSERT INTO c VALUES ");
-        for i in 1..=25 {
-            fill.push_str(&format!("({}, {}, {}),", i, i * 10, i * 1000));
-        }
-        fill.pop();
-        for sql in [
-            "ALTER DATABASE CURRENT SET READ_COMMITTED_SNAPSHOT ON",
-            "CREATE TABLE c (id INT NOT NULL PRIMARY KEY, v INT, w INT)",
-            fill.as_str(),
-            "CREATE INDEX ix_cv ON c (v) INCLUDE (v, w)",
-            // The open transaction updates only the INCLUDE column: the
-            // seek still finds the entry under v = 10, now carrying the NEW
-            // include bytes.
-            "BEGIN TRAN; UPDATE c SET w = 777 WHERE id = 1;",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut writer);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-
-        let select = |label: &str| {
-            let mut reader = TxnContext::default();
-            let outcome = execute_batch(&storage, "SELECT w FROM c WHERE v = 10", &mut reader);
-            assert!(outcome.error.is_none(), "{label}: {:?}", outcome.error);
-            match &outcome.results[0] {
-                StatementResult::Rows(rowset) => rowset.rows.clone(),
-                other => panic!("{label}: expected rows, got {other:?}"),
-            }
-        };
-
-        let covering_before = storage.covering_scans();
-        let rows = select("during the writer's transaction");
-        assert!(
-            storage.covering_scans() > covering_before,
-            "the read must have gone down the covering path for this to test anything"
-        );
-        assert_eq!(
-            rows,
-            vec![vec![Datum::Int(1000)]],
-            "the snapshot image, not the new include payload"
-        );
-
-        let outcome = execute_batch(&storage, "COMMIT", &mut writer);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        assert_eq!(rows_i32(&select("after the commit")), vec![777]);
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    fn rows_i32(rows: &[Vec<Datum>]) -> Vec<i32> {
-        rows.iter()
-            .map(|row| match row[0] {
-                Datum::Int(v) => v,
-                ref other => panic!("expected INT, got {other:?}"),
-            })
-            .collect()
-    }
-
-    /// Rollback unpublishes: a rolled-back transaction leaves no version
-    /// chains behind (pruning would otherwise treat its entries as an open
-    /// transaction's and pin them forever), and pruning drops the history of
-    /// committed transactions once no snapshot is live.
-    #[test]
-    fn rollback_unpublishes_and_prune_drops_settled_history() {
-        use crate::engine::{TxnContext, execute_batch};
-
-        let path = unique_temp_path("rcsi-prune");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        for sql in [
-            "ALTER DATABASE CURRENT SET READ_COMMITTED_SNAPSHOT ON",
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)",
-            "INSERT INTO t VALUES (1, 10), (2, 20)",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-
-        // The seed INSERTs published their own (committed) chains; settle
-        // them first so the rollback assertion sees only the rollback's work.
-        storage
-            .ensure_durable(storage.wal_tail())
-            .expect("durability");
-        storage.version_prune();
-        assert_eq!(storage.version_chain_count("t"), 0);
-
-        // Rolled back: the chains its statements published are reversed.
-        let outcome = execute_batch(
-            &storage,
-            "BEGIN TRAN; UPDATE t SET v = 99 WHERE id = 1; ROLLBACK",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        assert_eq!(
-            storage.version_chain_count("t"),
-            0,
-            "a rolled-back transaction leaves no chains"
-        );
-
-        // Committed: the chain exists until pruning decides nothing can need
-        // it (no live snapshot, commit durable).
-        let outcome = execute_batch(&storage, "UPDATE t SET v = 11 WHERE id = 1", &mut ctx);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        assert_eq!(storage.version_chain_count("t"), 1);
-        storage
-            .ensure_durable(storage.wal_tail())
-            .expect("durability");
-        storage.version_prune();
-        assert_eq!(
-            storage.version_chain_count("t"),
-            0,
-            "settled history is dropped by the maintenance prune"
-        );
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Version cleanup under load (Stage 13 exit): a live snapshot pins
-    /// exactly the history it may still read through sustained churn — and
-    /// keeps reading its own consistent view — while releasing it lets the
-    /// maintenance prune drop everything.
-    #[test]
-    fn version_cleanup_under_load_pins_then_drops_history() {
-        use crate::engine::{TxnContext, execute_batch};
-
-        let path = unique_temp_path("cleanup-load");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let mut seed = String::from("INSERT INTO t VALUES ");
-        for i in 0..100 {
-            seed.push_str(&format!("({i}, 0),"));
-        }
-        seed.pop();
-        for sql in [
-            "ALTER DATABASE CURRENT SET READ_COMMITTED_SNAPSHOT ON",
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)",
-            seed.as_str(),
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        storage
-            .ensure_durable(storage.wal_tail())
-            .expect("durability");
-        storage.version_prune();
-        assert_eq!(storage.version_chain_count("t"), 0, "settled baseline");
-
-        // A long-lived snapshot (an idle SNAPSHOT transaction's view) while
-        // a writer churns every row, five rounds, pruning between rounds.
-        let pinned = storage.capture_read_snapshot(None);
-        for round in 1..=5 {
-            let outcome = execute_batch(&storage, &format!("UPDATE t SET v = {round}"), &mut ctx);
-            assert!(outcome.error.is_none(), "{:?}", outcome.error);
-            storage
-                .ensure_durable(storage.wal_tail())
-                .expect("durability");
-            storage.version_prune();
-        }
-        assert_eq!(
-            storage.version_chain_count("t"),
-            100,
-            "the live snapshot pins one chain per churned row"
-        );
-        // The pinned view still reads its consistent state through all of it.
-        let rows = storage
-            .rel_scan_snapshot(
-                crate::relstore::catalog::DEFAULT_DATABASE_ID,
-                "t",
-                Some(&[1]),
-                pinned,
-            )
-            .expect("snapshot scan");
-        assert_eq!(rows.len(), 100);
-        assert!(
-            rows.iter().all(|r| r == &vec![Datum::Int(0)]),
-            "the snapshot sees the pre-churn value on every row"
-        );
-
-        // Released: the next prune drops the whole store.
-        storage.release_read_snapshot(pinned.seq);
-        storage.version_prune();
-        assert_eq!(
-            storage.version_chain_count("t"),
-            0,
-            "released history is dropped, the store is bounded"
-        );
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// SI review PoC: the 3960 auto-abort path must release the
-    /// transaction's snapshot registration - a leak would pin the prune
-    /// watermark forever. Observable through pruning: while the snapshot is
-    /// registered the conflicting chain must survive a prune; after the 3960
-    /// rolled the transaction back, the same prune must drop it.
-    #[test]
-    fn a_3960_abort_releases_the_snapshot_registration() {
-        use crate::engine::{TxnContext, execute_batch};
-
-        let path = unique_temp_path("si-3960-release");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut setup = TxnContext::default();
-        for sql in [
-            "ALTER DATABASE CURRENT SET ALLOW_SNAPSHOT_ISOLATION ON",
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)",
-            "INSERT INTO t VALUES (1, 10)",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut setup);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        storage
-            .ensure_durable(storage.wal_tail())
-            .expect("durability");
-        storage.version_prune();
-        assert_eq!(storage.version_chain_count("t"), 0);
-
-        // B: SNAPSHOT transaction, snapshot captured at first access.
-        let mut b = TxnContext::default();
-        let outcome = execute_batch(
-            &storage,
-            "SET TRANSACTION ISOLATION LEVEL SNAPSHOT; BEGIN TRAN; SELECT v FROM t",
-            &mut b,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-
-        // A: a conflicting committed write B's snapshot cannot see.
-        let mut a = TxnContext::default();
-        let outcome = execute_batch(&storage, "UPDATE t SET v = 99 WHERE id = 1", &mut a);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-
-        // While B's snapshot is registered, its history is pinned.
-        storage
-            .ensure_durable(storage.wal_tail())
-            .expect("durability");
-        storage.version_prune();
-        assert_eq!(
-            storage.version_chain_count("t"),
-            1,
-            "a registered snapshot pins the chain"
-        );
-
-        // B writes the same row: 3960, and the whole transaction (with its
-        // snapshot registration) must be gone.
-        let outcome = execute_batch(&storage, "UPDATE t SET v = 100 WHERE id = 1", &mut b);
-        assert_eq!(
-            outcome.error.as_ref().map(|e| e.number),
-            Some(3960),
-            "{:?}",
-            outcome.error
-        );
-
-        storage.version_prune();
-        assert_eq!(
-            storage.version_chain_count("t"),
-            0,
-            "the 3960 abort must release the snapshot registration"
-        );
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// The Stage 14 exit criterion: a 10 MiB value round-trips through an
-    /// overflow chain, survives a kill-and-reopen, and a mid-transaction
-    /// update of it recovers to the committed value.
-    #[test]
-    fn ten_mib_value_round_trips_and_survives_a_crash() {
-        use crate::engine::RpcParam;
-        use crate::engine::{
-            StatementResult, TxnContext, execute_batch, execute_batch_with_params,
-        };
-
-        let path = unique_temp_path("max-10mib");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        for sql in [
-            "CREATE TABLE big (id INT NOT NULL PRIMARY KEY, body NVARCHAR(MAX))",
-            "INSERT INTO big VALUES (1, N'seed')",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        // 10 MiB of UTF-16 payload = 5 * 1024 * 1024 characters.
-        let big: String = "abcdefgh".repeat(5 * 1024 * 1024 / 8);
-        assert_eq!(big.encode_utf16().count() * 2, 10 * 1024 * 1024);
-        let outcome = execute_batch_with_params(
-            &storage,
-            "UPDATE big SET body = @v WHERE id = 1",
-            &mut ctx,
-            &[RpcParam {
-                name: "@v".into(),
-                column_type: ColumnType::NVarCharMax,
-                value: crate::relstore::types::Datum::NVarChar(big.clone()),
-            }],
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-
-        let fetch = |storage: &Storage, ctx: &mut TxnContext| -> String {
-            let outcome = execute_batch(storage, "SELECT body FROM big WHERE id = 1", ctx);
-            assert!(outcome.error.is_none(), "{:?}", outcome.error);
-            match &outcome.results[0] {
-                StatementResult::Rows(rowset) => match &rowset.rows[0][0] {
-                    Datum::NVarChar(s) => s.clone(),
-                    other => panic!("expected NVARCHAR, got {other:?}"),
-                },
-                other => panic!("expected rows, got {other:?}"),
-            }
-        };
-        assert_eq!(
-            fetch(&storage, &mut ctx),
-            big,
-            "round-trip before the crash"
-        );
-
-        // An in-flight update dies with the crash; the committed 10 MiB
-        // value must recover intact.
-        let outcome = execute_batch(
-            &storage,
-            "BEGIN TRAN; UPDATE big SET body = N'doomed' WHERE id = 1;",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        let _ = ctx;
-        drop(storage);
-
-        let storage = Storage::open(path.clone()).expect("recovery");
-        let mut ctx = TxnContext::default();
-        let recovered = fetch(&storage, &mut ctx);
-        assert_eq!(recovered.len(), big.len(), "length after recovery");
-        assert_eq!(
-            recovered, big,
-            "the committed chain survives, the loser is undone"
-        );
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Versioning over (MAX): an RCSI reader sees the pre-update big value
-    /// through the version image's overflow REFERENCE (images carry raw row
-    /// bytes; chains are immutable and never freed, so the ref stays valid).
-    #[test]
-    fn rcsi_reads_the_old_big_value_through_the_image_reference() {
-        use crate::engine::{
-            RpcParam, StatementResult, TxnContext, execute_batch, execute_batch_with_params,
-        };
-
-        let path = unique_temp_path("max-rcsi");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut writer = TxnContext::default();
-        let big_old: String = "old-value".repeat(20_000); // ~180 KB
-        let big_new: String = "new-value".repeat(20_000);
-        for sql in [
-            "ALTER DATABASE CURRENT SET READ_COMMITTED_SNAPSHOT ON",
-            "CREATE TABLE big (id INT NOT NULL PRIMARY KEY, body NVARCHAR(MAX))",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut writer);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        let outcome = execute_batch_with_params(
-            &storage,
-            "INSERT INTO big VALUES (1, @v)",
-            &mut writer,
-            &[RpcParam {
-                name: "@v".into(),
-                column_type: ColumnType::NVarCharMax,
-                value: Datum::NVarChar(big_old.clone()),
-            }],
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-
-        // Writer holds an uncommitted update to the big value...
-        let outcome = execute_batch(&storage, "BEGIN TRAN", &mut writer);
-        assert!(outcome.error.is_none());
-        let outcome = execute_batch_with_params(
-            &storage,
-            "UPDATE big SET body = @v WHERE id = 1",
-            &mut writer,
-            &[RpcParam {
-                name: "@v".into(),
-                column_type: ColumnType::NVarCharMax,
-                value: Datum::NVarChar(big_new.clone()),
-            }],
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-
-        // ...and a snapshot reader gets the OLD value, resolved through the
-        // image's overflow reference.
-        let mut reader = TxnContext::default();
-        let outcome = execute_batch(&storage, "SELECT body FROM big WHERE id = 1", &mut reader);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => {
-                assert_eq!(rowset.rows[0][0], Datum::NVarChar(big_old.clone()));
-            }
-            other => panic!("expected rows, got {other:?}"),
-        }
-        let outcome = execute_batch(&storage, "COMMIT", &mut writer);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        let outcome = execute_batch(&storage, "SELECT body FROM big WHERE id = 1", &mut reader);
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => {
-                assert_eq!(rowset.rows[0][0], Datum::NVarChar(big_new.clone()));
-            }
-            other => panic!("expected rows, got {other:?}"),
-        }
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Review PoC: HEAP tables' version priors come from `heap.read_row`
-    /// pre-images. An RCSI reader must see the pre-update big value of a heap
-    /// row (resolved through the image's overflow reference), and a deleted
-    /// heap row must stay visible to the open snapshot.
-    #[test]
-    fn heap_rcsi_reads_old_big_value_through_preimage() {
-        use crate::engine::{
-            RpcParam, StatementResult, TxnContext, execute_batch, execute_batch_with_params,
-        };
-
-        let path = unique_temp_path("review-heap-rcsi");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut writer = TxnContext::default();
-        let big_old: String = "old-heap".repeat(10_000); // 80k chars -> chain
-        let big_new: String = "new-heap".repeat(10_000);
-        for sql in [
-            "ALTER DATABASE CURRENT SET READ_COMMITTED_SNAPSHOT ON",
-            // No PRIMARY KEY: a heap.
-            "CREATE TABLE hbig (id INT NOT NULL, body NVARCHAR(MAX))",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut writer);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        let outcome = execute_batch_with_params(
-            &storage,
-            "INSERT INTO hbig VALUES (1, @v)",
-            &mut writer,
-            &[RpcParam {
-                name: "@v".into(),
-                column_type: ColumnType::NVarCharMax,
-                value: Datum::NVarChar(big_old.clone()),
-            }],
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-
-        let fetch = |ctx: &mut TxnContext| -> Option<Datum> {
-            let outcome = execute_batch(&storage, "SELECT body FROM hbig WHERE id = 1", ctx);
-            assert!(outcome.error.is_none(), "{:?}", outcome.error);
-            match &outcome.results[0] {
-                StatementResult::Rows(rowset) => rowset.rows.first().map(|r| r[0].clone()),
-                other => panic!("expected rows, got {other:?}"),
-            }
-        };
-
-        // Uncommitted UPDATE: the reader sees the old value via the image.
-        let outcome = execute_batch(&storage, "BEGIN TRAN", &mut writer);
-        assert!(outcome.error.is_none());
-        let outcome = execute_batch_with_params(
-            &storage,
-            "UPDATE hbig SET body = @v WHERE id = 1",
-            &mut writer,
-            &[RpcParam {
-                name: "@v".into(),
-                column_type: ColumnType::NVarCharMax,
-                value: Datum::NVarChar(big_new.clone()),
-            }],
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        let mut reader = TxnContext::default();
-        assert_eq!(
-            fetch(&mut reader),
-            Some(Datum::NVarChar(big_old.clone())),
-            "heap RCSI reader must get the pre-update value"
-        );
-        let outcome = execute_batch(&storage, "COMMIT", &mut writer);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        assert_eq!(fetch(&mut reader), Some(Datum::NVarChar(big_new.clone())));
-
-        // Uncommitted DELETE: the reader still sees the (new) value.
-        let outcome = execute_batch(
-            &storage,
-            "BEGIN TRAN; DELETE FROM hbig WHERE id = 1;",
-            &mut writer,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        assert_eq!(
-            fetch(&mut reader),
-            Some(Datum::NVarChar(big_new.clone())),
-            "heap RCSI reader must see the row an open txn deleted"
-        );
-        let outcome = execute_batch(&storage, "COMMIT", &mut writer);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        assert_eq!(fetch(&mut reader), None);
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Review PoC: a heap row that has MOVED (forwarding stub at its home
-    /// RID) must still read correctly, version correctly (the pre-image is
-    /// read through the stub), and keep its overflow value.
-    #[test]
-    fn moved_heap_row_versions_and_resolves_through_the_stub() {
-        use crate::engine::{
-            RpcParam, StatementResult, TxnContext, execute_batch, execute_batch_with_params,
-        };
-
-        let path = unique_temp_path("review-heap-moved");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut writer = TxnContext::default();
-        let big_old: String = "moved-old".repeat(5_000); // 45k chars -> chain
-        let big_new: String = "moved-new".repeat(5_000);
-        for sql in [
-            "ALTER DATABASE CURRENT SET READ_COMMITTED_SNAPSHOT ON",
-            "CREATE TABLE hm (id INT NOT NULL, pad VARCHAR(3000), body NVARCHAR(MAX))",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut writer);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        // Row 1 small; row 2 fills most of the first heap page, so growing
-        // row 1's pad to 3000 bytes cannot fit and must move the row.
-        let outcome = execute_batch_with_params(
-            &storage,
-            "INSERT INTO hm VALUES (1, 'a', @v)",
-            &mut writer,
-            &[RpcParam {
-                name: "@v".into(),
-                column_type: ColumnType::NVarCharMax,
-                value: Datum::NVarChar(big_old.clone()),
-            }],
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        let filler = "f".repeat(3000);
-        let outcome = execute_batch(
-            &storage,
-            &format!("INSERT INTO hm VALUES (2, '{filler}', NULL)"),
-            &mut writer,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        let grown = "g".repeat(3000);
-        let outcome = execute_batch(
-            &storage,
-            &format!("UPDATE hm SET pad = '{grown}' WHERE id = 1"),
-            &mut writer,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-
-        // The moved row still resolves its chain.
-        let fetch_body = |ctx: &mut TxnContext| -> Option<Datum> {
-            let outcome = execute_batch(&storage, "SELECT body FROM hm WHERE id = 1", ctx);
-            assert!(outcome.error.is_none(), "{:?}", outcome.error);
-            match &outcome.results[0] {
-                StatementResult::Rows(rowset) => rowset.rows.first().map(|r| r[0].clone()),
-                other => panic!("expected rows, got {other:?}"),
-            }
-        };
-        assert_eq!(
-            fetch_body(&mut writer),
-            Some(Datum::NVarChar(big_old.clone())),
-            "moved row reads its chain"
-        );
-
-        // Version an update of the moved row: the prior must be read through
-        // the forwarding stub, and the reader must get the old big value.
-        let outcome = execute_batch(&storage, "BEGIN TRAN", &mut writer);
-        assert!(outcome.error.is_none());
-        let outcome = execute_batch_with_params(
-            &storage,
-            "UPDATE hm SET body = @v WHERE id = 1",
-            &mut writer,
-            &[RpcParam {
-                name: "@v".into(),
-                column_type: ColumnType::NVarCharMax,
-                value: Datum::NVarChar(big_new.clone()),
-            }],
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        let mut reader = TxnContext::default();
-        assert_eq!(
-            fetch_body(&mut reader),
-            Some(Datum::NVarChar(big_old.clone())),
-            "RCSI reader must get the pre-update value of a MOVED heap row"
-        );
-        // The pad read through the same image must be the grown one.
-        let outcome = execute_batch(&storage, "SELECT pad FROM hm WHERE id = 1", &mut reader);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => {
-                assert_eq!(rowset.rows[0][0], Datum::VarChar(grown.clone()));
-            }
-            other => panic!("expected rows, got {other:?}"),
-        }
-        let outcome = execute_batch(&storage, "COMMIT", &mut writer);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        assert_eq!(fetch_body(&mut reader), Some(Datum::NVarChar(big_new)));
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Review PoC: a statement that spills a chain and THEN fails (duplicate
-    /// key on a later row) must roll back cleanly — the chain leaks, the rows
-    /// do not land, the store stays usable, and recovery after a kill agrees.
-    #[test]
-    fn failed_statement_after_spill_rolls_back_cleanly() {
-        use crate::engine::{
-            RpcParam, StatementResult, TxnContext, execute_batch, execute_batch_with_params,
-        };
-
-        let path = unique_temp_path("review-spill-fail");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        for sql in [
-            "CREATE TABLE sf (id INT NOT NULL PRIMARY KEY, body NVARCHAR(MAX))",
-            "INSERT INTO sf VALUES (7, N'anchor')",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        let big: String = "spilled".repeat(10_000); // 70k chars -> chain
-        // Row 1 spills its chain, row 2 hits the duplicate key: the whole
-        // statement must fail and undo row 1.
-        let outcome = execute_batch_with_params(
-            &storage,
-            "INSERT INTO sf VALUES (1, @v), (7, N'dup')",
-            &mut ctx,
-            &[RpcParam {
-                name: "@v".into(),
-                column_type: ColumnType::NVarCharMax,
-                value: Datum::NVarChar(big.clone()),
-            }],
-        );
-        assert!(outcome.error.is_some(), "duplicate key must fail");
-
-        let count = |ctx: &mut TxnContext| -> Datum {
-            let outcome = execute_batch(&storage, "SELECT COUNT(*) FROM sf", ctx);
-            assert!(outcome.error.is_none(), "{:?}", outcome.error);
-            match &outcome.results[0] {
-                StatementResult::Rows(rowset) => rowset.rows[0][0].clone(),
-                other => panic!("expected rows, got {other:?}"),
-            }
-        };
-        assert_eq!(count(&mut ctx), Datum::BigInt(1), "only the anchor row");
-
-        // The store is still usable: the same big value inserts fine now.
-        let outcome = execute_batch_with_params(
-            &storage,
-            "INSERT INTO sf VALUES (1, @v)",
-            &mut ctx,
-            &[RpcParam {
-                name: "@v".into(),
-                column_type: ColumnType::NVarCharMax,
-                value: Datum::NVarChar(big.clone()),
-            }],
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        assert_eq!(count(&mut ctx), Datum::BigInt(2));
-
-        // Kill and reopen: recovery replays the leaked chain's images and
-        // the committed rows; the value survives.
-        drop(storage);
-        let storage = Storage::open(path.clone()).expect("recovery");
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(&storage, "SELECT body FROM sf WHERE id = 1", &mut ctx);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => {
-                assert_eq!(rowset.rows[0][0], Datum::NVarChar(big));
-            }
-            other => panic!("expected rows, got {other:?}"),
-        }
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Review PoC: ALTER TABLE ADD on a table with spilled (MAX) values.
-    /// The rewrite resolves every row and re-encodes it WITHOUT re-spilling,
-    /// so a chain value small enough for the row cap is silently re-inlined
-    /// (and must survive), while a big one fails the whole ALTER — which must
-    /// fail CLEANLY, leaving the table intact and the store usable.
-    #[test]
-    fn alter_add_column_respills_max_values() {
-        use crate::engine::{
-            RpcParam, StatementResult, TxnContext, execute_batch, execute_batch_with_params,
-        };
-
-        let path = unique_temp_path("review-alter-max");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-
-        // Case A: a 300-byte chain value fits back inline; ALTER succeeds.
-        let small_chain = "s".repeat(300); // VARCHAR: 300 bytes > 256 -> chain
-        for sql in [
-            "CREATE TABLE amax (id INT NOT NULL PRIMARY KEY, body VARCHAR(MAX))".to_string(),
-            format!("INSERT INTO amax VALUES (1, '{small_chain}')"),
-            "ALTER TABLE amax ADD extra INT NULL".to_string(),
-        ] {
-            let outcome = execute_batch(&storage, &sql, &mut ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        let outcome = execute_batch(&storage, "SELECT body, extra FROM amax", &mut ctx);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => {
-                assert_eq!(rowset.rows[0][0], Datum::VarChar(small_chain));
-                assert_eq!(rowset.rows[0][1], Datum::Null);
-            }
-            other => panic!("expected rows, got {other:?}"),
-        }
-
-        // Case B: a 10k value re-spills to a fresh chain during the ALTER's
-        // rewrite (the review found the original rewrite re-inlined and
-        // failed; the re-encode now runs inside the statement with
-        // spill_max_values, like every other write path).
-        let big: String = "b".repeat(10_000);
-        let outcome = execute_batch_with_params(
-            &storage,
-            "CREATE TABLE bmax (id INT NOT NULL PRIMARY KEY, body VARCHAR(MAX)); \
-             INSERT INTO bmax VALUES (1, @v);",
-            &mut ctx,
-            &[RpcParam {
-                name: "@v".into(),
-                column_type: ColumnType::VarCharMax,
-                value: Datum::VarChar(big.clone()),
-            }],
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        let outcome = execute_batch(&storage, "ALTER TABLE bmax ADD extra INT NULL", &mut ctx);
-        assert!(
-            outcome.error.is_none(),
-            "the rewrite must spill, not re-inline: {:?}",
-            outcome.error
-        );
-        let outcome = execute_batch(
-            &storage,
-            "SELECT body, extra FROM bmax WHERE id = 1",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => {
-                assert_eq!(rowset.rows[0][0], Datum::VarChar(big.clone()));
-                assert_eq!(rowset.rows[0][1], Datum::Null, "the frozen fill");
-            }
-            other => panic!("expected rows, got {other:?}"),
-        }
-        // ...and the widened row survives a reopen (the fresh chain is
-        // durable with the ALTER's statement).
-        drop(ctx);
-        drop(storage);
-        let storage = Storage::open(path.clone()).expect("reopen");
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(&storage, "SELECT body FROM bmax WHERE id = 1", &mut ctx);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => {
-                assert_eq!(rowset.rows[0][0], Datum::VarChar(big));
-            }
-            other => panic!("expected rows, got {other:?}"),
-        }
-        let outcome = execute_batch(&storage, "INSERT INTO bmax VALUES (2, 'ok', 5)", &mut ctx);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Review PoC: codec edges — empty string, the 256/257 inline threshold,
-    /// NULL, and a VARBINARY(MAX) value — all round-trip, including across a
-    /// reopen.
-    #[test]
-    fn max_codec_edges_round_trip() {
-        use crate::engine::{
-            RpcParam, StatementResult, TxnContext, execute_batch, execute_batch_with_params,
-        };
-
-        let path = unique_temp_path("review-max-edges");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let at_threshold = "t".repeat(256); // inline boundary
-        let over_threshold = "u".repeat(257); // first chained length
-        for sql in [
-            "CREATE TABLE edges (id INT NOT NULL PRIMARY KEY, v VARCHAR(MAX), b VARBINARY(MAX))"
-                .to_string(),
-            "INSERT INTO edges VALUES (1, '', NULL)".to_string(),
-            format!("INSERT INTO edges VALUES (2, '{at_threshold}', NULL)"),
-            format!("INSERT INTO edges VALUES (3, '{over_threshold}', NULL)"),
-            "INSERT INTO edges VALUES (4, NULL, NULL)".to_string(),
-        ] {
-            let outcome = execute_batch(&storage, &sql, &mut ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        let blob = vec![0xABu8; 300]; // > 256 -> chain
-        let outcome = execute_batch_with_params(
-            &storage,
-            "INSERT INTO edges VALUES (5, NULL, @b)",
-            &mut ctx,
-            &[RpcParam {
-                name: "@b".into(),
-                column_type: ColumnType::VarBinaryMax,
-                value: Datum::VarBinary(blob.clone()),
-            }],
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-
-        let check = |storage: &Storage, ctx: &mut TxnContext| {
-            let outcome = execute_batch(storage, "SELECT v, b FROM edges ORDER BY id", ctx);
-            assert!(outcome.error.is_none(), "{:?}", outcome.error);
-            match &outcome.results[0] {
-                StatementResult::Rows(rowset) => {
-                    assert_eq!(rowset.rows[0][0], Datum::VarChar(String::new()), "empty");
-                    assert_eq!(rowset.rows[1][0], Datum::VarChar(at_threshold.clone()));
-                    assert_eq!(rowset.rows[2][0], Datum::VarChar(over_threshold.clone()));
-                    assert_eq!(rowset.rows[3][0], Datum::Null);
-                    assert_eq!(rowset.rows[4][1], Datum::VarBinary(blob.clone()));
-                }
-                other => panic!("expected rows, got {other:?}"),
-            }
-        };
-        check(&storage, &mut ctx);
-        drop(storage);
-        let storage = Storage::open(path.clone()).expect("reopen");
-        let mut ctx = TxnContext::default();
-        check(&storage, &mut ctx);
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Offline growth (Stage 14): the data region extends, everything in it
-    /// survives, and the grown space is allocatable. Includes a second grow
-    /// (the re-run shape an interrupted grow needs).
-    #[test]
-    fn offline_grow_extends_the_data_region() {
-        use crate::engine::{StatementResult, TxnContext, execute_batch};
-
-        let path = unique_temp_path("grow");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        for sql in [
-            "CREATE TABLE g (id INT NOT NULL PRIMARY KEY, v NVARCHAR(MAX))",
-            "INSERT INTO g VALUES (1, N'before growth')",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        let old_pages = storage.lock().layout.data_size / PAGE_SIZE as u64;
-        drop(ctx);
-        drop(storage);
-
-        let new_pages = Storage::grow_data_region(&path, 1).expect("grow");
-        assert_eq!(
-            new_pages,
-            old_pages + (1u64 << 30) / PAGE_SIZE as u64,
-            "one GiB of new data pages"
-        );
-
-        let storage = Storage::open(path.clone()).expect("reopen after grow");
-        assert_eq!(
-            storage.lock().layout.data_size / PAGE_SIZE as u64,
-            new_pages
-        );
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(&storage, "SELECT v FROM g WHERE id = 1", &mut ctx);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => {
-                assert_eq!(rowset.rows[0][0], Datum::NVarChar("before growth".into()));
-            }
-            other => panic!("expected rows, got {other:?}"),
-        }
-        let outcome = execute_batch(&storage, "INSERT INTO g VALUES (2, N'after')", &mut ctx);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        drop(ctx);
-        drop(storage);
-
-        // Growing again works (the re-run an interrupted grow performs).
-        let newer = Storage::grow_data_region(&path, 1).expect("second grow");
-        assert_eq!(newer, new_pages + (1u64 << 30) / PAGE_SIZE as u64);
-        let storage = Storage::open(path.clone()).expect("reopen after second grow");
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(&storage, "SELECT COUNT(*) FROM g", &mut ctx);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// The safe minimum: a delta below the tail span is refused with the
-    /// minimum named; nothing is touched.
-    #[test]
-    fn grow_refuses_below_the_safe_minimum() {
-        let path = unique_temp_path("grow-min");
-        // 4 GiB file: the tail regions span ~1.16 GiB, so +1 GiB is unsafe.
-        let mut opts = test_storage_options();
-        opts.size_gib = 4;
-        let storage = Storage::create(path.clone(), opts).expect("create");
-        drop(storage);
-        let err = Storage::grow_data_region(&path, 1).expect_err("must refuse");
-        assert!(
-            err.to_string().contains("safe minimum"),
-            "names the floor: {err}"
-        );
-        let storage = Storage::open(path.clone()).expect("file untouched");
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Growth with a pending WAL (a crash-interrupted transaction) is safe:
-    /// the WAL sits before the data region and replays against the moved
-    /// bitmap exactly as it would have.
-    #[test]
-    fn grow_with_pending_wal_recovers_cleanly() {
-        use crate::engine::{StatementResult, TxnContext, execute_batch};
-
-        let path = unique_temp_path("grow-wal");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        for sql in [
-            "CREATE TABLE g (id INT NOT NULL PRIMARY KEY, v INT)",
-            "INSERT INTO g VALUES (1, 10)",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        // An open transaction dies with the "crash" (drop without commit).
-        let outcome = execute_batch(
-            &storage,
-            "BEGIN TRAN; UPDATE g SET v = 999 WHERE id = 1;",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        drop(ctx);
-        drop(storage);
-
-        Storage::grow_data_region(&path, 1).expect("grow with pending WAL");
-        let storage = Storage::open(path.clone()).expect("recovery after grow");
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(&storage, "SELECT v FROM g WHERE id = 1", &mut ctx);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => {
-                assert_eq!(rowset.rows[0][0], Datum::Int(10), "the loser is undone");
-            }
-            other => panic!("expected rows, got {other:?}"),
-        }
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// The advisory file lock fences grow against a running server: grow
-    /// refuses while the store is open, and works after it closes. (flock is
-    /// per open file description, so two opens in one process conflict too.)
-    #[test]
-    fn grow_refuses_while_the_store_is_open() {
-        let path = unique_temp_path("grow-flock");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let err = Storage::grow_data_region(&path, 2).expect_err("must refuse while open");
-        assert!(
-            err.to_string()
-                .contains("locked by another TruthDB process"),
-            "names the lock: {err}"
-        );
-        drop(storage);
-        Storage::grow_data_region(&path, 2).expect("grow after close");
-        let storage = Storage::open(path.clone()).expect("reopen");
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// REVIEW POC: a crash after the extension writes but BEFORE the header
-    /// stamp leaves a longer file under the old header. The file must open
-    /// under the old layout, and a re-run of the grow must complete it.
-    /// Simulated by saving the original header page before the grow and
-    /// writing it back afterwards.
-    #[test]
-    fn grow_crash_before_header_stamp_is_recoverable() {
-        use crate::engine::{StatementResult, TxnContext, execute_batch};
-
-        let path = unique_temp_path("grow-crash");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        for sql in [
-            "CREATE TABLE g (id INT NOT NULL PRIMARY KEY, v INT)",
-            "INSERT INTO g VALUES (1, 42)",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        // A real checkpoint makes the descriptor and bitmap copies
-        // load-bearing (fresh files have neither on disk).
-        storage
-            .write_checkpoint(b"crash-window-probe", 1, 2, 1)
-            .expect("checkpoint");
-        drop(ctx);
-        drop(storage);
-
-        // Save the pre-grow header page (the commit point the "crash" loses).
-        let old_header = {
-            let mut f = std::fs::File::open(&path).expect("open for header save");
-            let mut buf = vec![0u8; FILE_HEADER_SIZE];
-            f.read_exact(&mut buf).expect("read header");
-            buf
-        };
-
-        let old_pages = {
-            let storage = Storage::open(path.clone()).expect("preflight open");
-            let pages = storage.lock().layout.data_size / PAGE_SIZE as u64;
-            drop(storage);
-            pages
-        };
-
-        Storage::grow_data_region(&path, 1).expect("grow");
-
-        // Crash simulation: the extension writes are durable, the header
-        // stamp is not.
-        {
-            let mut f = std::fs::OpenOptions::new()
-                .write(true)
-                .open(&path)
-                .expect("open for header restore");
-            f.seek(SeekFrom::Start(0)).expect("seek");
-            f.write_all(&old_header).expect("restore old header");
-            f.sync_all().expect("sync");
-        }
-
-        // The old layout must be fully valid: data readable, snapshot loads.
-        let storage = Storage::open(path.clone()).expect("open under the OLD header");
-        assert_eq!(
-            storage.lock().layout.data_size / PAGE_SIZE as u64,
-            old_pages,
-            "still the old layout"
-        );
-        let snap = storage
-            .load_snapshot()
-            .expect("load snapshot under old header")
-            .expect("snapshot present");
-        assert_eq!(snap.data, b"crash-window-probe");
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(&storage, "SELECT v FROM g WHERE id = 1", &mut ctx);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => {
-                assert_eq!(rowset.rows[0][0], Datum::Int(42));
-            }
-            other => panic!("expected rows, got {other:?}"),
-        }
-        drop(ctx);
-        drop(storage);
-
-        // Re-running the grow completes it.
-        let new_pages = Storage::grow_data_region(&path, 1).expect("re-run grow");
-        assert_eq!(new_pages, old_pages + (1u64 << 30) / PAGE_SIZE as u64);
-        let storage = Storage::open(path.clone()).expect("open after completed grow");
-        assert_eq!(
-            storage.lock().layout.data_size / PAGE_SIZE as u64,
-            new_pages
-        );
-        let snap = storage
-            .load_snapshot()
-            .expect("load snapshot after grow")
-            .expect("snapshot survived");
-        assert_eq!(snap.data, b"crash-window-probe");
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(&storage, "INSERT INTO g VALUES (2, 43)", &mut ctx);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// REVIEW POC (teeth): the checkpointed search snapshot must survive a
-    /// grow — the descriptor pages are the only pointer to it, and they move.
-    /// Fails if the grow skips the descriptor copy.
-    #[test]
-    fn grow_preserves_the_checkpointed_snapshot() {
-        let path = unique_temp_path("grow-snap");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        storage
-            .write_checkpoint(b"survives-the-move", 1, 2, 1)
-            .expect("checkpoint");
-        drop(storage);
-
-        Storage::grow_data_region(&path, 1).expect("grow");
-
-        let storage = Storage::open(path.clone()).expect("reopen");
-        let snap = storage
-            .load_snapshot()
-            .expect("load")
-            .expect("snapshot survived the grow");
-        assert_eq!(snap.data, b"survives-the-move");
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// REVIEW POC (teeth): after a checkpoint the persisted bitmap is the
-    /// ONLY record of table extents (the WAL head has advanced past their
-    /// alloc records). If the grow loses the bitmap, a reopen sees every page
-    /// free and new allocations clobber existing tables. Fails if the grow
-    /// skips the bitmap copy.
-    #[test]
-    fn grow_preserves_the_persisted_allocator_bitmap() {
-        use crate::engine::{StatementResult, TxnContext, execute_batch};
-
-        let path = unique_temp_path("grow-bitmap");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        for sql in [
-            "CREATE TABLE t1 (id INT NOT NULL PRIMARY KEY, v NVARCHAR(60))",
-            "INSERT INTO t1 VALUES (1, N'pre-grow row')",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        drop(ctx);
-        // The checkpoint persists the bitmap and advances the WAL head past
-        // t1's extent-alloc records.
-        storage
-            .write_checkpoint(b"bitmap-probe", 1, 2, 1)
-            .expect("checkpoint");
-        drop(storage);
-
-        Storage::grow_data_region(&path, 1).expect("grow");
-
-        // Reopen and allocate heavily; with a lost bitmap these allocations
-        // land on t1's pages.
-        let storage = Storage::open(path.clone()).expect("reopen");
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(
-            &storage,
-            "CREATE TABLE t2 (id INT NOT NULL PRIMARY KEY, v NVARCHAR(200))",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        for batch in 0..20 {
-            let mut sql = String::from("INSERT INTO t2 VALUES ");
-            for i in 0..50 {
-                if i > 0 {
-                    sql.push(',');
-                }
-                let id = batch * 50 + i;
-                sql.push_str(&format!("({id}, N'filler row {id} {}')", "x".repeat(120)));
-            }
-            let outcome = execute_batch(&storage, &sql, &mut ctx);
-            assert!(
-                outcome.error.is_none(),
-                "batch {batch}: {:?}",
-                outcome.error
-            );
-        }
-        let outcome = execute_batch(&storage, "SELECT v FROM t1 WHERE id = 1", &mut ctx);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => {
-                assert_eq!(
-                    rowset.rows[0][0],
-                    Datum::NVarChar("pre-grow row".into()),
-                    "t1 must not be clobbered by post-grow allocations"
-                );
-            }
-            other => panic!("expected rows, got {other:?}"),
-        }
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Publishing versions changes nothing about crash recovery: the store
-    /// (chains and commit table) is memory-only, so a kill-and-reopen with
-    /// RCSI on recovers exactly the committed state, options intact, chains
-    /// empty.
-    #[test]
-    fn rcsi_survives_a_crash_with_clean_recovery() {
-        use crate::engine::{StatementResult, TxnContext, execute_batch};
-
-        let path = unique_temp_path("rcsi-crash");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        for sql in [
-            "ALTER DATABASE CURRENT SET READ_COMMITTED_SNAPSHOT ON",
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)",
-            "INSERT INTO t VALUES (1, 10), (2, 20)",
-            "UPDATE t SET v = 11 WHERE id = 1",
-        ] {
-            let outcome = execute_batch(&storage, sql, &mut ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        }
-        // An open transaction with published versions dies with the crash.
-        let outcome = execute_batch(
-            &storage,
-            "BEGIN TRAN; UPDATE t SET v = 999 WHERE id = 2;",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        // Kill without checkpoint: drop the handle mid-transaction.
-        let _ = ctx;
-        drop(storage);
-
-        let storage = Storage::open(path.clone()).expect("recovery");
-        assert!(storage.rcsi_enabled());
-        assert_eq!(
-            storage.version_chain_count("t"),
-            0,
-            "chains do not survive a restart"
-        );
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(&storage, "SELECT v FROM t ORDER BY id", &mut ctx);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => assert_eq!(
-                rows_i32(&rowset.rows),
-                vec![11, 20],
-                "committed wins, the in-flight update is undone"
-            ),
-            other => panic!("expected rows, got {other:?}"),
-        }
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Row-lock escalation (Stage 12), pinned at its actual threshold: a
-    /// statement naming more than 1000 row keys takes ONE table lock instead
-    /// of flooding the lock table; at or below the threshold it takes row
-    /// locks. (The plan sketched ~5000; 1000 is the shipped value — this
-    /// test is the record of that divergence.)
-    #[test]
-    fn row_locks_escalate_past_the_threshold() {
-        use crate::engine::{Isolation, TxnContext, execute_batch};
-        use crate::lock::{LockMode, Resource};
-
-        let path = unique_temp_path("escalation");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(
-            &storage,
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-
-        let insert = |n: usize| {
-            let tuples: Vec<String> = (0..n).map(|i| format!("({i}, 0)")).collect();
-            format!("INSERT INTO t VALUES {}", tuples.join(", "))
-        };
-
-        // At the threshold: per-row locks (plus the table intent).
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            &insert(1000),
-            Isolation::ReadCommitted,
-        );
-        let rows = needs
-            .iter()
-            .filter(|(r, _)| matches!(r, Resource::Row(_, _)))
-            .count();
-        assert_eq!(rows, 1000, "at the threshold every key gets a row lock");
-        assert!(
-            !needs
-                .iter()
-                .any(|(r, m)| matches!(r, Resource::Table(_)) && *m == LockMode::Exclusive),
-            "no table X below escalation: {:?}",
-            needs.len()
-        );
-
-        // A single statement past the threshold: the per-statement cap
-        // declines to enumerate 1001 row hashes and the INSERT falls back to
-        // one table-exclusive lock. (Reachable since the node budget became
-        // per-expression — a 1001-tuple INSERT parses now.)
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            &insert(1001),
-            Isolation::ReadCommitted,
-        );
-        assert!(
-            needs
-                .iter()
-                .any(|(r, m)| matches!(r, Resource::Table(_)) && *m == LockMode::Exclusive),
-            "a single over-threshold statement takes table X: {needs:?}"
-        );
-        assert_eq!(
-            needs
-                .iter()
-                .filter(|(r, _)| matches!(r, Resource::Row(_, _)))
-                .count(),
-            0
-        );
-
-        // Past it — summed across the WHOLE BATCH: a 1000-tuple INSERT plus
-        // 20 point DELETEs on DISTINCT keys wants 1020 row locks on one
-        // table (the needs map dedups by key hash, so overlapping keys would
-        // not count twice), and the batch-level pass replaces them all with
-        // one table-exclusive lock.
-        let deletes: Vec<String> = (2000..2020)
-            .map(|i| format!("DELETE FROM t WHERE id = {i}"))
-            .collect();
-        let over = format!("{}; {}", insert(1000), deletes.join("; "));
-        let needs = crate::engine::analyze_locks(
-            &storage,
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            &over,
-            Isolation::ReadCommitted,
-        );
-        assert!(
-            needs
-                .iter()
-                .any(|(r, m)| matches!(r, Resource::Table(_)) && *m == LockMode::Exclusive),
-            "past the threshold the statement takes table X: {needs:?}"
-        );
-        assert_eq!(
-            needs
-                .iter()
-                .filter(|(r, _)| matches!(r, Resource::Row(_, _)))
-                .count(),
-            0,
-            "row locks are replaced, not added to"
-        );
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// EXEC scope semantics, pinned: outer variables are restored after the
-    /// inner batch (a regression previously went green), and SET options
-    /// revert at scope exit as SQL Server reverts them — an inner
-    /// `SET XACT_ABORT ON` must not doom the outer batch's transaction.
-    #[test]
-    fn exec_scope_restores_variables_and_set_options() {
-        use crate::engine::{StatementResult, TxnContext, execute_batch};
-
-        let path = unique_temp_path("exec-scope");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(
-            &storage,
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-
-        // Inner @o shadows; the outer @o is intact after the EXEC returns.
-        let outcome = execute_batch(
-            &storage,
-            "DECLARE @o INT = 1; EXEC sp_executesql N'DECLARE @o INT = 99; SELECT @o AS inner_o'; SELECT @o AS outer_o",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        let values: Vec<i64> = outcome
-            .results
-            .iter()
-            .filter_map(|r| match r {
-                StatementResult::Rows(rs) => match rs.rows[0][0] {
-                    Datum::Int(v) => Some(v as i64),
-                    Datum::BigInt(v) => Some(v),
-                    _ => None,
-                },
-                _ => None,
-            })
-            .collect();
-        assert_eq!(values, [99, 1], "{:?}", outcome.results);
-
-        // An inner SET XACT_ABORT ON reverts at EXEC exit: the later duplicate
-        // key fails its own statement but the batch continues and commits.
-        let outcome = execute_batch(
-            &storage,
-            "BEGIN TRANSACTION; INSERT INTO t VALUES (1); \
-             EXEC sp_executesql N'SET XACT_ABORT ON'; \
-             INSERT INTO t VALUES (1); INSERT INTO t VALUES (2); COMMIT",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_some(), "the dup insert reports its error");
-        let outcome = execute_batch(&storage, "SELECT COUNT(*) FROM t", &mut ctx);
-        match &outcome.results[0] {
-            StatementResult::Rows(rs) => assert_eq!(
-                rs.rows[0][0],
-                Datum::BigInt(2),
-                "XACT_ABORT must revert at scope exit; the transaction commits"
-            ),
-            other => panic!("expected rows, got {other:?}"),
-        }
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    fn unique_temp_path(label: &str) -> PathBuf {
-        let mut path = std::env::temp_dir();
-        let nanos = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .expect("system time before unix epoch")
-            .as_nanos();
-        path.push(format!("truthdb-storage-{label}-{nanos}.db"));
-        path
-    }
-
-    fn create_small(path: &Path) -> Storage {
-        Storage::create_with_wal_bounds(
-            path.to_path_buf(),
-            test_storage_options(),
-            TEST_WAL_BYTES,
-            TEST_WAL_BYTES,
-        )
-        .expect("create storage")
-    }
-
-    /// A covering seek (every read column INCLUDEd) answers from the index
-    /// leaves alone — the counter proves the covering path ran — and returns
-    /// exactly what a table scan returns, original case and NULLs included.
-    #[test]
-    fn a_covering_seek_answers_from_the_index_alone_and_matches_a_scan() {
-        use crate::engine::{StatementResult, TxnContext, execute_batch};
-
-        let path = unique_temp_path("include-covering");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let setup = execute_batch(
-            &storage,
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, email VARCHAR(40) NOT NULL, v INT); \
-             CREATE INDEX ix ON t (email) INCLUDE (email, v); \
-             INSERT INTO t VALUES (1, 'a@x.com', 10), (2, 'B@X.com', NULL), (3, 'c@x.com', 30)",
-            &mut ctx,
-        );
-        assert!(setup.error.is_none(), "{:?}", setup.error);
-
-        let rows_of = |outcome: &crate::engine::BatchOutcome| match &outcome.results[0] {
-            StatementResult::Rows(rowset) => rowset.rows.clone(),
-            other => panic!("expected rows, got {other:?}"),
-        };
-
-        // Sought case-insensitively, answered with the stored ORIGINAL value.
-        let covered = execute_batch(
-            &storage,
-            "SELECT email, v FROM t WHERE email = 'b@x.com'",
-            &mut ctx,
-        );
-        assert!(covered.error.is_none(), "{:?}", covered.error);
-        assert_eq!(storage.covering_scans(), 1, "the covering path answered");
-        let covered = rows_of(&covered);
-        assert_eq!(
-            covered,
-            vec![vec![Datum::VarChar("B@X.com".into()), Datum::Null]]
-        );
-
-        // A/B: without the index the same query scans — identical rows.
-        let dropped = execute_batch(&storage, "DROP INDEX ix ON t", &mut ctx);
-        assert!(dropped.error.is_none(), "{:?}", dropped.error);
-        let scanned = execute_batch(
-            &storage,
-            "SELECT email, v FROM t WHERE email = 'b@x.com'",
-            &mut ctx,
-        );
-        assert!(scanned.error.is_none(), "{:?}", scanned.error);
-        assert_eq!(rows_of(&scanned), covered, "covering == scan");
-        assert_eq!(storage.covering_scans(), 1, "the scan path took over");
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// The streamed input path: a filtered aggregate over a plain base table
-    /// pulls the scan slice by slice through the WHERE — it must never drain
-    /// the scan whole (that is what the join operators do, and what the old
-    /// path did for every shape). Pinned by the materialization counter: the
-    /// streamed query performs zero whole-scan drains; a join performs one
-    /// per scanned input.
-    #[test]
-    fn a_filtered_aggregate_streams_its_input_instead_of_materializing() {
-        use crate::engine::{StatementResult, TxnContext, execute_batch};
-
-        let path = unique_temp_path("stream-input");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let setup = execute_batch(
-            &storage,
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT NOT NULL)",
-            &mut ctx,
-        );
-        assert!(setup.error.is_none(), "{:?}", setup.error);
-        for chunk in (1..=3000).collect::<Vec<i64>>().chunks(500) {
-            let values: Vec<String> = chunk.iter().map(|i| format!("({i}, {})", i % 7)).collect();
-            let outcome = execute_batch(
-                &storage,
-                &format!("INSERT INTO t VALUES {}", values.join(", ")),
-                &mut ctx,
-            );
-            assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        }
-
-        // An aggregate with a WHERE over 3000 rows (three scan slices): the
-        // input streams; nothing drains the scan whole.
-        let outcome = execute_batch(
-            &storage,
-            "SELECT COUNT(*), SUM(v) FROM t WHERE v > 3",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        assert_eq!(
-            storage.scan_materializations(),
-            0,
-            "the filtered aggregate's input streamed"
-        );
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => {
-                // v cycles 1..7 over 3000 rows minus the id=1 seed row's v=1:
-                // v in {4,5,6} appears 428|429 times; exact values pin the walk.
-                assert_eq!(rowset.rows.len(), 1);
-            }
-            other => panic!("expected rows, got {other:?}"),
-        }
-
-        // A join materializes only its BUILD side (the counter's positive
-        // control); the probe side streams. INNER probes from the left, so the
-        // right input is the one drained.
-        let outcome = execute_batch(
-            &storage,
-            "SELECT COUNT(*) FROM t a JOIN t b ON a.id = b.id WHERE a.v > 5",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        assert_eq!(
-            storage.scan_materializations(),
-            1,
-            "an inner hash join materializes only its build side"
-        );
-
-        // RIGHT reverses orientation: the probe is the right input, the left
-        // input is built — still exactly one materialization.
-        let outcome = execute_batch(
-            &storage,
-            "SELECT COUNT(*) FROM t a RIGHT JOIN t b ON a.id = b.id",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        assert_eq!(
-            storage.scan_materializations(),
-            2,
-            "a right join materializes only its (left) build side"
-        );
-
-        // The nested loop (no equi key) streams its probe side too. A small
-        // table keeps the O(n·m) loop cheap; its base scan is still lazy.
-        let setup = execute_batch(
-            &storage,
-            "CREATE TABLE s (id INT NOT NULL PRIMARY KEY); INSERT INTO s VALUES (1), (2), (3), (4)",
-            &mut ctx,
-        );
-        assert!(setup.error.is_none(), "{:?}", setup.error);
-        let outcome = execute_batch(
-            &storage,
-            "SELECT COUNT(*) FROM s a JOIN s b ON a.id < b.id",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => assert_eq!(rowset.rows[0][0], Datum::BigInt(6)),
-            other => panic!("expected rows, got {other:?}"),
-        }
-        assert_eq!(
-            storage.scan_materializations(),
-            3,
-            "a nested-loop join materializes only its build side"
-        );
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// The grace-hash spill path partitions the probe side straight off the
-    /// scan stream: still exactly one materialization (the build side), and
-    /// the same result as the in-memory join.
-    #[test]
-    fn a_spilling_join_streams_its_probe_side_into_partitions() {
-        use crate::engine::{StatementResult, TxnContext, execute_batch, set_test_sort_budget};
-
-        let path = unique_temp_path("grace-probe-stream");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let setup = execute_batch(
-            &storage,
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT NOT NULL)",
-            &mut ctx,
-        );
-        assert!(setup.error.is_none(), "{:?}", setup.error);
-        for chunk in (1..=2000).collect::<Vec<i64>>().chunks(500) {
-            let values: Vec<String> = chunk.iter().map(|i| format!("({i}, {})", i % 7)).collect();
-            let outcome = execute_batch(
-                &storage,
-                &format!("INSERT INTO t VALUES {}", values.join(", ")),
-                &mut ctx,
-            );
-            assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        }
-
-        set_test_sort_budget(Some(4000));
-        let outcome = execute_batch(
-            &storage,
-            "SELECT COUNT(*) FROM t a LEFT JOIN t b ON a.id = b.id AND b.v = 3",
-            &mut ctx,
-        );
-        set_test_sort_budget(None);
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        match &outcome.results[0] {
-            // Every a-row appears once: matched where b.v = 3, null-extended
-            // otherwise — 2000 either way.
-            StatementResult::Rows(rowset) => assert_eq!(rowset.rows[0][0], Datum::BigInt(2000)),
-            other => panic!("expected rows, got {other:?}"),
-        }
-        assert_eq!(
-            storage.scan_materializations(),
-            1,
-            "the spilling join materialized only its build side"
-        );
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// The row counter tracks every DML shape through the SQL layer — and
-    /// because it is an ordinary transactional page op, statement atomicity,
-    /// savepoints, transaction rollback and crash recovery all keep it exact
-    /// without counter-specific recovery code.
-    #[test]
-    fn row_counts_track_dml_transactions_and_recovery() {
-        use crate::engine::{TxnContext, execute_batch};
-
-        let path = unique_temp_path("row-count");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let run = |storage: &Storage, ctx: &mut TxnContext, sql: &str| {
-            let outcome = execute_batch(storage, sql, ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-        };
-
-        run(
-            &storage,
-            &mut ctx,
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)",
-        );
-        assert_eq!(
-            storage.rel_row_count(crate::relstore::catalog::DEFAULT_DATABASE_ID, "t"),
-            Some(0)
-        );
-
-        run(
-            &storage,
-            &mut ctx,
-            "INSERT INTO t VALUES (1, 10), (2, 20), (3, 30)",
-        );
-        assert_eq!(
-            storage.rel_row_count(crate::relstore::catalog::DEFAULT_DATABASE_ID, "t"),
-            Some(3)
-        );
-
-        run(&storage, &mut ctx, "DELETE FROM t WHERE id = 3");
-        run(&storage, &mut ctx, "UPDATE t SET v = 99 WHERE id = 1");
-        assert_eq!(
-            storage.rel_row_count(crate::relstore::catalog::DEFAULT_DATABASE_ID, "t"),
-            Some(2),
-            "delete -1, update 0"
-        );
-
-        // A failing multi-row statement (duplicate key on its last row) is
-        // atomic: no rows land, and neither does its count.
-        let dup = execute_batch(&storage, "INSERT INTO t VALUES (5, 1), (1, 1)", &mut ctx);
-        assert!(dup.error.is_some(), "duplicate key must fail");
-        assert_eq!(
-            storage.rel_row_count(crate::relstore::catalog::DEFAULT_DATABASE_ID, "t"),
-            Some(2)
-        );
-
-        // Transaction rollback restores the count with the rows.
-        run(
-            &storage,
-            &mut ctx,
-            "BEGIN TRANSACTION; INSERT INTO t VALUES (10, 1), (11, 1)",
-        );
-        assert_eq!(
-            storage.rel_row_count(crate::relstore::catalog::DEFAULT_DATABASE_ID, "t"),
-            Some(4),
-            "in-flight rows count"
-        );
-        run(&storage, &mut ctx, "ROLLBACK");
-        assert_eq!(
-            storage.rel_row_count(crate::relstore::catalog::DEFAULT_DATABASE_ID, "t"),
-            Some(2)
-        );
-
-        // A savepoint rollback restores exactly the statements behind it.
-        run(
-            &storage,
-            &mut ctx,
-            "BEGIN TRANSACTION; INSERT INTO t VALUES (20, 1); SAVE TRANSACTION sp; \
-             INSERT INTO t VALUES (21, 1); ROLLBACK TRANSACTION sp; COMMIT",
-        );
-        assert_eq!(
-            storage.rel_row_count(crate::relstore::catalog::DEFAULT_DATABASE_ID, "t"),
-            Some(3)
-        );
-
-        // Crash (no checkpoint, pool never flushed): recovery replays the ops,
-        // counter page included.
-        drop(storage);
-        let storage = Storage::open(path.clone()).expect("reopen");
-        assert_eq!(
-            storage.rel_row_count(crate::relstore::catalog::DEFAULT_DATABASE_ID, "t"),
-            Some(3),
-            "count survives recovery"
-        );
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Rows of a transaction still open at the crash are undone by recovery —
-    /// and so is their count.
-    #[test]
-    fn an_uncommitted_transactions_rows_are_uncounted_after_crash() {
-        use crate::engine::{TxnContext, execute_batch};
-
-        let path = unique_temp_path("row-count-crash");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let setup = execute_batch(
-            &storage,
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY); INSERT INTO t VALUES (1), (2)",
-            &mut ctx,
-        );
-        assert!(setup.error.is_none(), "{:?}", setup.error);
-
-        let mut open_txn = TxnContext::default();
-        let pending = execute_batch(
-            &storage,
-            "BEGIN TRANSACTION; INSERT INTO t VALUES (10), (11), (12)",
-            &mut open_txn,
-        );
-        assert!(pending.error.is_none(), "{:?}", pending.error);
-        assert_eq!(
-            storage.rel_row_count(crate::relstore::catalog::DEFAULT_DATABASE_ID, "t"),
-            Some(5),
-            "in-flight rows count"
-        );
-        drop(storage); // crash with the transaction open
-
-        let storage = Storage::open(path.clone()).expect("reopen");
-        assert_eq!(
-            storage.rel_row_count(crate::relstore::catalog::DEFAULT_DATABASE_ID, "t"),
-            Some(2),
-            "the loser transaction's rows and count are both undone"
-        );
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// "Row counts as tie-breakers only": a table at or under the tiny
-    /// threshold plans its seek as the scan it ties with, grows into the seek
-    /// past the threshold — and a covering seek keeps its win at any size,
-    /// since it reads less than the table either way.
-    #[test]
-    fn a_tiny_table_scans_until_it_grows_into_its_seek() {
-        use crate::engine::{StatementResult, TxnContext, execute_batch};
-
-        let path = unique_temp_path("row-count-tiebreak");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let run = |storage: &Storage, ctx: &mut TxnContext, sql: &str| {
-            let outcome = execute_batch(storage, sql, ctx);
-            assert!(outcome.error.is_none(), "{sql}: {:?}", outcome.error);
-            outcome
-        };
-        let plan_of = |storage: &Storage, ctx: &mut TxnContext, sql: &str| -> String {
-            let outcome = run(storage, ctx, &format!("SET SHOWPLAN_TEXT ON; {sql}"));
-            let mut ctx2 = TxnContext::default();
-            let _ = execute_batch(storage, "SET SHOWPLAN_TEXT OFF", &mut ctx2);
-            match &outcome.results[1] {
-                StatementResult::Rows(rowset) => rowset
-                    .rows
-                    .iter()
-                    .map(|r| format!("{:?}", r[0]))
-                    .collect::<Vec<_>>()
-                    .join("\n"),
-                other => panic!("expected plan rows, got {other:?}"),
-            }
-        };
-
-        run(
-            &storage,
-            &mut ctx,
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, a INT, v INT); \
-             CREATE INDEX ix_a ON t (a); \
-             CREATE INDEX ix_cover ON t (v) INCLUDE (v, id); \
-             INSERT INTO t VALUES (1, 10, 5), (2, 20, 6), (3, 30, 7)",
-        );
-
-        // Tiny: the non-covering seek ties with the scan; the tie goes to the
-        // scan. The covering seek still wins — it reads less than the table.
-        let tiny = plan_of(&storage, &mut ctx, "SELECT id FROM t WHERE a = 20");
-        assert!(tiny.contains("Table Scan"), "tiny table scans: {tiny}");
-        let covering = plan_of(&storage, &mut ctx, "SELECT v, id FROM t WHERE v = 6");
-        assert!(
-            covering.contains("Index Seek (covering)"),
-            "covering exempt from the tie-break: {covering}"
-        );
-
-        // Past the threshold the same query seeks.
-        let mut pad = String::from("INSERT INTO t VALUES (100, 900, 900)");
-        for i in 1..20 {
-            pad.push_str(&format!(", ({}, 900, 900)", 100 + i));
-        }
-        run(&storage, &mut ctx, &pad);
-        let grown = plan_of(&storage, &mut ctx, "SELECT id FROM t WHERE a = 20");
-        assert!(
-            grown.contains("Index Seek") && grown.contains("Key Lookup"),
-            "grown table seeks: {grown}"
-        );
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// A covering index wins its tie against an equal-scoring non-INCLUDE
-    /// index — the "add a covering index to an existing database" workflow.
-    /// Coverage breaks equality ties only: it never outranks a fully-matched
-    /// UNIQUE seek (one row plus one lookup beats a covering scan).
-    #[test]
-    fn a_covering_index_wins_the_tie_against_an_older_plain_index() {
-        use crate::engine::{TxnContext, execute_batch};
-
-        let path = unique_temp_path("include-tiebreak");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        // The plain index is created FIRST, so a first-wins tie keeps it.
-        let setup = execute_batch(
-            &storage,
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, email VARCHAR(40) NOT NULL, v INT); \
-             CREATE INDEX ix ON t (email); \
-             CREATE INDEX ix2 ON t (email) INCLUDE (email, v); \
-             INSERT INTO t VALUES (1, 'a@x.com', 10)",
-            &mut ctx,
-        );
-        assert!(setup.error.is_none(), "{:?}", setup.error);
-        let outcome = execute_batch(
-            &storage,
-            "SELECT email, v FROM t WHERE email = 'a@x.com'",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        assert_eq!(
-            storage.covering_scans(),
-            1,
-            "the covering index wins the equality tie"
-        );
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// CREATE INDEX with a column that does not exist on the table reports
-    /// SQL Server's 1911 (not the generic 207) — for the key list and the
-    /// INCLUDE list alike; a duplicate INCLUDE column reports 1909.
-    #[test]
-    fn create_index_errors_carry_sql_server_numbers() {
-        use crate::engine::{TxnContext, execute_batch};
-
-        let path = unique_temp_path("include-errors");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let setup = execute_batch(
-            &storage,
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, email VARCHAR(40) NOT NULL)",
-            &mut ctx,
-        );
-        assert!(setup.error.is_none(), "{:?}", setup.error);
-
-        let cases = [
-            ("CREATE INDEX ix ON t (nope)", 1911),
-            ("CREATE INDEX ix ON t (email) INCLUDE (nope)", 1911),
-            ("CREATE INDEX ix ON t (email) INCLUDE (id, id)", 1909),
-        ];
-        for (sql, number) in cases {
-            let outcome = execute_batch(&storage, sql, &mut ctx);
-            assert_eq!(
-                outcome.error.as_ref().map(|e| e.number),
-                Some(number),
-                "{sql}: {:?}",
-                outcome.error
-            );
-        }
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Included leaf values follow UPDATE and DELETE, and survive a restart
-    /// (the include list rides the catalog; the leaf format rides the pages).
-    #[test]
-    fn included_values_survive_update_delete_and_restart() {
-        use crate::engine::{StatementResult, TxnContext, execute_batch};
-
-        let path = unique_temp_path("include-restart");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let setup = execute_batch(
-            &storage,
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, email VARCHAR(40) NOT NULL, v INT); \
-             CREATE INDEX ix ON t (email) INCLUDE (email, v); \
-             INSERT INTO t VALUES (1, 'a@x.com', 10), (2, 'b@x.com', 20); \
-             UPDATE t SET v = 99 WHERE id = 1; \
-             DELETE FROM t WHERE id = 2",
-            &mut ctx,
-        );
-        assert!(setup.error.is_none(), "{:?}", setup.error);
-        drop(storage);
-
-        let storage = Storage::open(path.clone()).expect("reopen");
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(
-            &storage,
-            "SELECT email, v FROM t WHERE email = 'a@x.com'",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        assert_eq!(storage.covering_scans(), 1, "covering after reopen");
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => assert_eq!(
-                rowset.rows,
-                vec![vec![Datum::VarChar("a@x.com".into()), Datum::Int(99)]],
-                "the UPDATE reached the leaf; the DELETEd row is gone"
-            ),
-            other => panic!("expected rows, got {other:?}"),
-        }
-        let gone = execute_batch(
-            &storage,
-            "SELECT email, v FROM t WHERE email = 'b@x.com'",
-            &mut ctx,
-        );
-        match &gone.results[0] {
-            StatementResult::Rows(rowset) => assert!(rowset.rows.is_empty()),
-            other => panic!("expected rows, got {other:?}"),
-        }
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// A query reading a column that is NOT included falls back to the key
-    /// lookup — through the INCLUDE index's length-prefixed leaf value, whose
-    /// `Locator::Key` payload would be swallowed by the old bare decode.
-    /// SHOWPLAN tells the two apart: covering has no Key Lookup line.
-    #[test]
-    fn a_non_covering_read_on_an_include_index_still_finds_rows() {
-        use crate::engine::{StatementResult, TxnContext, execute_batch};
-
-        let path = unique_temp_path("include-fallback");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut ctx = TxnContext::default();
-        let setup = execute_batch(
-            &storage,
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, email VARCHAR(40) NOT NULL, name VARCHAR(20)); \
-             CREATE INDEX ix ON t (email) INCLUDE (email, id); \
-             INSERT INTO t VALUES (1, 'a@x.com', 'Alice')",
-            &mut ctx,
-        );
-        assert!(setup.error.is_none(), "{:?}", setup.error);
-        // Pad past the tiny-table tie-break: a <= 16-row table plans as a
-        // scan, and this test is about the seek's two plan shapes.
-        let mut pad = String::from("INSERT INTO t VALUES (100, 'z0@x.com', 'p')");
-        for i in 1..20 {
-            pad.push_str(&format!(", ({}, 'z{i}@x.com', 'p')", 100 + i));
-        }
-        let setup = execute_batch(&storage, &pad, &mut ctx);
-        assert!(setup.error.is_none(), "{:?}", setup.error);
-
-        // `name` is not included: the seek fetches the base row by PK key.
-        let outcome = execute_batch(
-            &storage,
-            "SELECT name FROM t WHERE email = 'a@x.com'",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        assert_eq!(storage.covering_scans(), 0, "not covering");
-        match &outcome.results[0] {
-            StatementResult::Rows(rowset) => {
-                assert_eq!(rowset.rows, vec![vec![Datum::VarChar("Alice".into())]]);
-            }
-            other => panic!("expected rows, got {other:?}"),
-        }
-
-        // SHOWPLAN: the covering shape has no Key Lookup; the fallback does.
-        let plans = execute_batch(
-            &storage,
-            "SET SHOWPLAN_TEXT ON; \
-             SELECT id FROM t WHERE email = 'a@x.com'; \
-             SELECT name FROM t WHERE email = 'a@x.com'",
-            &mut ctx,
-        );
-        assert!(plans.error.is_none(), "{:?}", plans.error);
-        let lines_of = |result: &StatementResult| -> Vec<String> {
-            match result {
-                StatementResult::Rows(rowset) => {
-                    rowset.rows.iter().map(|r| format!("{:?}", r[0])).collect()
-                }
-                other => panic!("expected plan rows, got {other:?}"),
-            }
-        };
-        let covering = lines_of(&plans.results[1]).join("\n");
-        assert!(
-            covering.contains("Index Seek (covering)") && !covering.contains("Key Lookup"),
-            "covering plan: {covering}"
-        );
-        let lookup = lines_of(&plans.results[2]).join("\n");
-        assert!(
-            lookup.contains("Key Lookup"),
-            "non-covering plan keeps the lookup: {lookup}"
-        );
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// The DONE acknowledging an autocommit write is never emitted before the
-    /// write's commit record is fsync-durable: the executor's deferred DONEs
-    /// flush — fsyncing first — before the next result set opens, so a client
-    /// can never act on an acknowledgment a crash then silently revokes.
-    #[test]
-    fn a_commit_acknowledgement_never_precedes_its_fsync() {
-        use crate::engine::{
-            BatchEmitter, ResultColumn, TxnContext, execute_batch, execute_batch_streamed,
-        };
-        use crate::relstore::types::Datum;
-
-        /// Logs each emitted event alongside the fsync count at that moment.
-        struct Probe<'a> {
-            storage: &'a Storage,
-            log: Vec<(&'static str, u64)>,
-        }
-        impl Probe<'_> {
-            fn note(&mut self, what: &'static str) {
-                self.log.push((what, self.storage.group_commit_fsyncs()));
-            }
-        }
-        impl BatchEmitter for Probe<'_> {
-            fn columns(&mut self, _columns: Vec<ResultColumn>) {
-                self.note("columns");
-            }
-            fn rows(&mut self, _rows: Vec<Vec<Datum>>) {
-                self.note("rows");
-            }
-            fn statement_done(
-                &mut self,
-                _count: Option<u64>,
-                _in_transaction: bool,
-                _command: crate::engine::DoneCommand,
-            ) {
-                self.note("done");
-            }
-            fn statement_aborted(&mut self, _in_transaction: bool) {
-                self.note("aborted");
-            }
-        }
-
-        let path = unique_temp_path("stream-durability");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut setup = TxnContext::default();
-        let create = execute_batch(&storage, "CREATE TABLE t (v INT NOT NULL)", &mut setup);
-        assert!(create.error.is_none(), "create table: {:?}", create.error);
-        let baseline = storage.group_commit_fsyncs();
-
-        let mut ctx = TxnContext::default();
-        let mut probe = Probe {
-            storage: &storage,
-            log: Vec::new(),
-        };
-        let error = execute_batch_streamed(
-            &storage,
-            "INSERT INTO t VALUES (1); SELECT v FROM t",
-            &mut ctx,
-            &[],
-            &mut probe,
-        );
-        assert!(error.is_none(), "{error:?}");
-        // The INSERT's DONE comes first — already past its fsync — and only
-        // then does the SELECT's result set open.
-        assert_eq!(
-            probe.log.first(),
-            Some(&("done", baseline + 1)),
-            "the write's acknowledgment waits for its fsync: {:?}",
-            probe.log
-        );
-        assert_eq!(
-            probe.log.get(1).map(|(what, _)| *what),
-            Some("columns"),
-            "the rowset opens after the acknowledgment: {:?}",
-            probe.log
-        );
-        // One fsync total: the mid-batch flush covered the batch's only commit.
-        assert_eq!(storage.group_commit_fsyncs(), baseline + 1);
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// An identity value reserved by an in-transaction INSERT (a mini-commit,
-    /// durable independently of the transaction) must not stream to the client
-    /// before its reservation is fsynced: a value the client has seen must
-    /// never be reissued after a crash, and it escapes through the *rows* of a
-    /// following SELECT, not through any commit-acknowledging DONE. This is
-    /// why the mid-batch flush gates on the kind-based `committed` flag rather
-    /// than on "some DONE acknowledges a commit".
-    #[test]
-    fn an_identity_value_never_streams_before_its_reservation_fsync() {
-        use crate::engine::{
-            BatchEmitter, ResultColumn, TxnContext, execute_batch, execute_batch_streamed,
-        };
-        use crate::relstore::types::Datum;
-
-        struct Probe<'a> {
-            storage: &'a Storage,
-            log: Vec<(&'static str, u64)>,
-        }
-        impl Probe<'_> {
-            fn note(&mut self, what: &'static str) {
-                self.log.push((what, self.storage.group_commit_fsyncs()));
-            }
-        }
-        impl BatchEmitter for Probe<'_> {
-            fn columns(&mut self, _columns: Vec<ResultColumn>) {
-                self.note("columns");
-            }
-            fn rows(&mut self, _rows: Vec<Vec<Datum>>) {
-                self.note("rows");
-            }
-            fn statement_done(
-                &mut self,
-                _count: Option<u64>,
-                _in_transaction: bool,
-                _command: crate::engine::DoneCommand,
-            ) {
-                self.note("done");
-            }
-            fn statement_aborted(&mut self, _in_transaction: bool) {
-                self.note("aborted");
-            }
-        }
-
-        let path = unique_temp_path("stream-identity");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut setup = TxnContext::default();
-        let create = execute_batch(
-            &storage,
-            "CREATE TABLE t (id INT IDENTITY(1,1) PRIMARY KEY, v INT NOT NULL)",
-            &mut setup,
-        );
-        assert!(create.error.is_none(), "create table: {:?}", create.error);
-        let baseline = storage.group_commit_fsyncs();
-
-        let mut ctx = TxnContext::default();
-        let mut probe = Probe {
-            storage: &storage,
-            log: Vec::new(),
-        };
-        // The INSERT's own DONE promises nothing durable (the transaction is
-        // open), but its identity reservation is already a mini-commit — and
-        // the SELECT's rows carry the reserved value out of the server.
-        let error = execute_batch_streamed(
-            &storage,
-            "BEGIN TRANSACTION; INSERT INTO t (v) VALUES (10); SELECT id FROM t; ROLLBACK",
-            &mut ctx,
-            &[],
-            &mut probe,
-        );
-        assert!(error.is_none(), "{error:?}");
-        for (what, fsyncs) in &probe.log {
-            if *what == "columns" || *what == "rows" {
-                assert!(
-                    *fsyncs > baseline,
-                    "the rowset streamed before the reservation's fsync: {:?}",
-                    probe.log
-                );
-            }
-        }
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// A batch of autocommit writes with nothing to stream between them still
-    /// coalesces to a single fsync at the end of the batch: the DONEs are
-    /// deferred to that durability point rather than each buying an fsync.
-    #[test]
-    fn a_write_only_batch_still_fsyncs_once() {
-        use crate::engine::{TxnContext, execute_batch};
-
-        let path = unique_temp_path("stream-one-fsync");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let mut setup = TxnContext::default();
-        let create = execute_batch(&storage, "CREATE TABLE t (v INT NOT NULL)", &mut setup);
-        assert!(create.error.is_none(), "create table: {:?}", create.error);
-        let baseline = storage.group_commit_fsyncs();
-
-        let mut ctx = TxnContext::default();
-        let outcome = execute_batch(
-            &storage,
-            "INSERT INTO t VALUES (1); INSERT INTO t VALUES (2); INSERT INTO t VALUES (3)",
-            &mut ctx,
-        );
-        assert!(outcome.error.is_none(), "{:?}", outcome.error);
-        assert_eq!(
-            storage.group_commit_fsyncs(),
-            baseline + 1,
-            "three autocommit writes, one batch-end fsync"
-        );
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Group commit: many transactions whose commit records are already in the
-    /// WAL, all waiting for durability up to the same point, are made durable by
-    /// a single log-writer fsync. Deterministic (independent of fsync latency):
-    /// the commit records are appended WITHOUT fsyncing first — `rel_insert`
-    /// appends its commit record but does not force the log — then every
-    /// committer waits on the same tail, so exactly one fsync serves them all.
-    #[test]
-    fn group_commit_coalesces_many_commits_into_one_fsync() {
-        use crate::engine::{TxnContext, execute_batch};
-        use std::sync::Arc;
-
-        const THREADS: usize = 16;
-
-        let path = unique_temp_path("group-commit");
-        let storage =
-            Arc::new(Storage::create(path.clone(), test_storage_options()).expect("create"));
-
-        let mut setup = TxnContext::default();
-        let create = execute_batch(&storage, "CREATE TABLE t (v INT NOT NULL)", &mut setup);
-        assert!(create.error.is_none(), "create table: {:?}", create.error);
-        let baseline = storage.group_commit_fsyncs();
-
-        // Raw autocommit inserts append a commit record each with `sync=false`
-        // and never call `ensure_durable`, so the WAL tail advances while
-        // `flushed` stays put — nothing fsyncs.
-        for i in 0..THREADS {
-            storage
-                .rel_insert(
-                    crate::relstore::catalog::DEFAULT_DATABASE_ID,
-                    "t",
-                    vec![Datum::Int(i as i32)],
-                )
-                .expect("insert");
-        }
-        let target = storage.wal_tail();
-        assert_eq!(
-            storage.group_commit_fsyncs(),
-            baseline,
-            "appending commit records must not fsync"
-        );
-
-        // Every committer waits for durability up to the same tail; one fsync
-        // covers them all.
-        let mut handles = Vec::new();
-        for _ in 0..THREADS {
-            let storage = Arc::clone(&storage);
-            handles.push(std::thread::spawn(move || {
-                storage.ensure_durable(target).expect("durable")
-            }));
-        }
-        for handle in handles {
-            handle.join().expect("thread panicked");
-        }
-
-        let fsyncs = storage.group_commit_fsyncs() - baseline;
-        assert!(
-            (1..=2).contains(&fsyncs),
-            "{THREADS} commits should coalesce into a single fsync, got {fsyncs}"
-        );
-
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// An identity value is consumed permanently (SQL Server semantics), so its
-    /// reservation — a mini-commit made even inside an open transaction — must be
-    /// fsynced. Regression: group commit skipped `ensure_durable` for a batch
-    /// whose only durable effect was the identity reservation (an INSERT inside
-    /// a still-open transaction), so a crash would revert and reuse the value.
-    #[test]
-    fn identity_reservation_is_made_durable_even_inside_a_transaction() {
-        use crate::engine::{TxnContext, execute_batch};
-
-        let path = unique_temp_path("identity-durable");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-
-        let mut ctx = TxnContext::default();
-        let create = execute_batch(
-            &storage,
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY IDENTITY(1,1), v INT NOT NULL)",
-            &mut ctx,
-        );
-        assert!(create.error.is_none(), "create: {:?}", create.error);
-
-        let before = storage.group_commit_fsyncs();
-        // INSERT inside an open transaction: no row commits (the COMMIT is a
-        // later batch), but the identity reservation does and must be fsynced.
-        let out = execute_batch(
-            &storage,
-            "BEGIN TRAN; INSERT INTO t (v) VALUES (1)",
-            &mut ctx,
-        );
-        assert!(out.error.is_none(), "insert: {:?}", out.error);
-        assert!(
-            storage.group_commit_fsyncs() > before,
-            "identity reservation inside a transaction must be made durable"
-        );
-
-        let _ = execute_batch(&storage, "ROLLBACK", &mut ctx);
-        drop(storage);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// Buffered write into the file (test-side corruption / fixtures). The
-    /// sync makes it visible to subsequent O_DIRECT reads.
-    fn overwrite_bytes(path: &Path, offset: u64, bytes: &[u8]) {
-        let mut file = std::fs::OpenOptions::new()
-            .write(true)
-            .open(path)
-            .expect("open for corruption");
-        file.seek(SeekFrom::Start(offset)).expect("seek");
-        file.write_all(bytes).expect("write");
-        file.sync_all().expect("sync");
-    }
-
-    fn read_bytes(path: &Path, offset: u64, len: usize) -> Vec<u8> {
-        let mut file = std::fs::File::open(path).expect("open for read");
-        file.seek(SeekFrom::Start(offset)).expect("seek");
-        let mut buf = vec![0u8; len];
-        file.read_exact(&mut buf).expect("read");
-        buf
-    }
-
-    fn append_search_entry(storage: &mut Storage, seq_no: u64, payload: &[u8]) -> u64 {
-        storage
-            .append_wal_entry(WAL_ENTRY_TYPE_RECORD, 1, seq_no, payload)
-            .expect("append wal entry")
-    }
-
-    /// Reads both superblocks from disk and returns the active (highest
-    /// valid generation) one.
-    fn read_active_superblock(path: &Path, layout: &StorageLayout) -> Superblock {
-        let read_sb = |offset: u64| -> Option<Superblock> {
-            let bytes = read_bytes(path, offset, SUPERBLOCK_SIZE);
-            let sb = Superblock::from_le_bytes(bytes.as_slice().try_into().unwrap());
-            (sb.checksum == sb.compute_checksum()).then_some(sb)
-        };
-        let a = read_sb(layout.superblock_a_offset);
-        let b = read_sb(layout.superblock_b_offset);
-        match (a, b) {
-            (Some(a), Some(b)) => {
-                if b.generation > a.generation {
-                    b
-                } else {
-                    a
-                }
-            }
-            (Some(a), None) => a,
-            (None, Some(b)) => b,
-            (None, None) => panic!("no valid superblock"),
-        }
-    }
-
-    fn test_layout() -> StorageLayout {
-        compute_layout(test_storage_options(), TEST_WAL_BYTES, TEST_WAL_BYTES).expect("layout")
-    }
-
-    /// Writes a v1-format file: v1 header, superblocks, optional snapshot in
-    /// half-slot 0, WAL entries with v1 stamping (`logical_ts = seq_no`) and
-    /// garbage in the allocator bitmap (v1 half-slot bookkeeping the upgrade
-    /// must discard).
-    fn write_v1_fixture(
-        path: &Path,
-        wal_events: &[(u64, Vec<u8>)],
-        snapshot: Option<(&[u8], u64, u64, u64)>,
-    ) -> StorageLayout {
-        let layout = test_layout();
-        let file = std::fs::OpenOptions::new()
-            .write(true)
-            .create_new(true)
-            .open(path)
-            .expect("create fixture");
-        file.set_len(layout.total_size).expect("set_len");
-        drop(file);
-
-        let mut header = FileHeader::default();
-        header.version = FILE_VERSION_V1;
-        header.superblock_a_offset = layout.superblock_a_offset;
-        header.superblock_b_offset = layout.superblock_b_offset;
-        header.wal_offset = layout.wal_offset;
-        header.wal_size = layout.wal_size;
-        header.data_offset = layout.data_offset;
-        header.data_size = layout.data_size;
-        header.metadata_offset = layout.metadata_offset;
-        header.metadata_size = layout.metadata_size;
-        header.allocator_offset = layout.allocator_offset;
-        header.allocator_size = layout.allocator_size;
-        header.snapshot_offset = layout.snapshot_offset;
-        header.snapshot_size = layout.snapshot_size;
-        header.reserved_offset = layout.reserved_offset;
-        header.reserved_size = layout.reserved_size;
-        header.header_checksum = header.compute_checksum();
-        overwrite_bytes(path, 0, &header.to_le_bytes_with_checksum());
-
-        // WAL entries, sequentially from ring position 0, v1 stamping.
-        let mut tail = 0u64;
-        let mut last_seq = 0u64;
-        for (seq_no, payload) in wal_events {
-            let padded = wal_entry_padded_len(payload.len());
-            let crc = wal_payload_crc(payload);
-            let entry_header = WalEntryHeader::new(
-                WAL_ENTRY_TYPE_RECORD,
-                1,
-                payload.len() as u32,
-                *seq_no,
-                *seq_no, // v1: logical_ts carried the engine seq
-                crc,
-            );
-            let footer = WalEntryFooter {
-                payload_len: payload.len() as u32,
-                payload_crc: crc,
-            };
-            let mut bytes = Vec::with_capacity(padded);
-            bytes.extend_from_slice(&entry_header.to_le_bytes());
-            bytes.extend_from_slice(payload);
-            bytes.extend_from_slice(&footer.to_le_bytes());
-            bytes.resize(padded, 0);
-            overwrite_bytes(path, layout.wal_offset + tail, &bytes);
-            tail += padded as u64;
-            last_seq = *seq_no;
-        }
-
-        // Snapshot in v1 half-slot 0 (start of the data region).
-        if let Some((data, checkpoint_seq, next_seq_no, next_doc_id)) = snapshot {
-            overwrite_bytes(path, layout.data_offset, data);
-            let mut desc = SnapshotDescriptor::default();
-            desc.generation = 1;
-            desc.slot = 0;
-            desc.checkpoint_seq = checkpoint_seq;
-            desc.data_offset = layout.data_offset;
-            desc.data_len = data.len() as u64;
-            desc.data_checksum = xxh64(data, 0);
-            desc.next_seq_no = next_seq_no;
-            desc.next_doc_id = next_doc_id;
-            desc.checksum = desc.compute_checksum();
-            overwrite_bytes(
-                path,
-                layout.snapshot_offset,
-                &desc.to_le_bytes_with_checksum(),
-            );
-        }
-
-        // v1 half-slot allocator garbage the upgrade must wipe.
-        overwrite_bytes(path, layout.allocator_offset, &[0xFF; 512]);
-
-        let mut sb_a = Superblock::default();
-        sb_a.generation = 2;
-        sb_a.active = SUPERBLOCK_ACTIVE_A;
-        sb_a.wal_tail = tail;
-        sb_a.last_committed_seq = last_seq;
-        sb_a.checksum = sb_a.compute_checksum();
-        overwrite_bytes(
-            path,
-            layout.superblock_a_offset,
-            &sb_a.to_le_bytes_with_checksum(),
-        );
-        let mut sb_b = Superblock::default();
-        sb_b.active = SUPERBLOCK_ACTIVE_B;
-        sb_b.checksum = sb_b.compute_checksum();
-        overwrite_bytes(
-            path,
-            layout.superblock_b_offset,
-            &sb_b.to_le_bytes_with_checksum(),
-        );
-        layout
-    }
-
-    #[test]
-    fn v1_fixture_upgrades_in_place_and_preserves_state() {
-        let path = unique_temp_path("v1-upgrade");
-        let snapshot_data = b"v1-snapshot-payload".as_slice();
-        let events = vec![(6u64, vec![1u8; 100]), (7u64, vec![2u8; 50])];
-        write_v1_fixture(&path, &events, Some((snapshot_data, 5, 8, 3)));
-
-        let mut storage = Storage::open(path.clone()).expect("open v1 file");
-
-        // Snapshot survives the upgrade.
-        let snapshot = storage
-            .load_snapshot()
-            .expect("load snapshot")
-            .expect("snapshot present");
-        assert_eq!(snapshot.data, snapshot_data);
-        assert_eq!(snapshot.checkpoint_seq, 5);
-        assert_eq!(snapshot.next_seq_no, 8);
-        assert_eq!(snapshot.next_doc_id, 3);
-
-        // WAL entries survive and replay in order.
-        let records = storage.replay_wal_entries().expect("replay");
-        assert_eq!(records.len(), 2);
-        assert_eq!(records[0].seq_no, 6);
-        assert_eq!(records[0].payload, vec![1u8; 100]);
-        assert_eq!(records[1].seq_no, 7);
-        assert_eq!(records[1].payload, vec![2u8; 50]);
-
-        // The allocator was rebuilt: only the live snapshot extent is
-        // allocated; the v1 half-slot garbage is gone.
-        assert!(storage.is_page_allocated(0), "snapshot extent must be live");
-        assert!(
-            !storage.is_page_allocated(1),
-            "pages past the snapshot extent must be free"
-        );
-        assert!(
-            !storage.is_page_allocated(100),
-            "v1 bitmap garbage must have been wiped"
-        );
-
-        // New work post-upgrade: append + checkpoint + reopen.
-        append_search_entry(&mut storage, 8, b"post-upgrade");
-        storage
-            .write_checkpoint(b"v2-snapshot", 8, 9, 4)
-            .expect("checkpoint");
-        drop(storage);
-
-        // On-disk version is now v2; upgraded file opens cleanly.
-        let header_bytes = read_bytes(&path, 0, FILE_HEADER_SIZE);
-        let header = FileHeader::from_le_bytes(header_bytes.as_slice().try_into().unwrap());
-        assert_eq!(header.version, FILE_VERSION);
-        assert_eq!(header.header_checksum, header.compute_checksum());
-
-        let mut storage = Storage::open(path.clone()).expect("reopen upgraded");
-        let snapshot = storage
-            .load_snapshot()
-            .expect("load")
-            .expect("second snapshot");
-        assert_eq!(snapshot.data, b"v2-snapshot");
-        assert!(
-            storage.replay_wal_entries().expect("replay").is_empty(),
-            "checkpoint reclaimed the wal"
-        );
-        // The upgraded snapshot's extent was freed once the v2 one became
-        // durable; page 0 belonged to the v1 snapshot.
-        assert!(!storage.is_page_allocated(0));
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-    }
-
-    #[test]
-    fn torn_tail_is_stopped_at_and_healed_by_whole_page_flush() {
-        let path = unique_temp_path("torn-tail");
-        let layout = test_layout();
-        let mut storage = create_small(&path);
-        let payload_a = vec![0xAAu8; 100];
-        append_search_entry(&mut storage, 1, &payload_a);
-        drop(storage); // crash: superblock still says tail = 0
-
-        // Simulate a torn write of a follow-up entry: garbage on the tail
-        // page right after entry A.
-        let entry_a_len = wal_entry_padded_len(payload_a.len()) as u64;
-        overwrite_bytes(&path, layout.wal_offset + entry_a_len, &[0x5Au8; 200]);
-
-        // Recovery must stop at the garbage and keep A.
-        let mut storage = Storage::open(path.clone()).expect("reopen after tear");
-        let records = storage.replay_wal_entries().expect("replay");
-        assert_eq!(records.len(), 1, "only entry A must survive the torn tail");
-        assert_eq!(records[0].payload, payload_a);
-
-        // The next append rewrites the whole tail page from memory, healing
-        // the torn bytes: B lands exactly where the garbage was.
-        let payload_b = vec![0xBBu8; 60];
-        append_search_entry(&mut storage, 2, &payload_b);
-        drop(storage);
-
-        let mut storage = Storage::open(path.clone()).expect("reopen after heal");
-        let records = storage.replay_wal_entries().expect("replay");
-        assert_eq!(records.len(), 2);
-        assert_eq!(records[0].payload, payload_a);
-        assert_eq!(records[1].payload, payload_b);
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-    }
-
-    #[test]
-    fn superblock_is_lazy_and_forward_scan_recovers_past_it() {
-        let path = unique_temp_path("lazy-superblock");
-        let layout = test_layout();
-        let mut storage = create_small(&path);
-        storage.lock().wal.set_superblock_interval(400);
-
-        // 152 bytes per entry: the cadence write fires once, on entry 3
-        // (456 bytes appended), and entry 4 stays past the recorded tail.
-        let payloads: Vec<Vec<u8>> = (0..4u8).map(|i| vec![i + 1; 100]).collect();
-        let mut lsns = Vec::new();
-        for (i, payload) in payloads.iter().enumerate() {
-            lsns.push(append_search_entry(&mut storage, i as u64 + 1, payload));
-        }
-        drop(storage); // crash without checkpoint
-
-        // The on-disk superblock lags the true tail (laziness) but is not 0
-        // (the cadence rewrite fired).
-        let sb = read_active_superblock(&path, &layout);
-        let true_tail = lsns[3] + wal_entry_padded_len(payloads[3].len()) as u64;
-        assert!(sb.wal_tail > 0, "cadence superblock write must have fired");
-        assert!(
-            sb.wal_tail < true_tail,
-            "superblock tail {} must lag the true tail {true_tail}",
-            sb.wal_tail
-        );
-
-        // Recovery scans forward past the stale superblock tail and finds
-        // every entry.
-        let mut storage = Storage::open(path.clone()).expect("reopen");
-        let records = storage.replay_wal_entries().expect("replay");
-        assert_eq!(records.len(), 4, "forward scan must recover all entries");
-        for (record, payload) in records.iter().zip(&payloads) {
-            assert_eq!(&record.payload, payload);
-        }
-        // LSN self-identity stamping.
-        for (record, lsn) in records.iter().zip(&lsns) {
-            assert_eq!(record.logical_ts, *lsn);
-        }
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-    }
-
-    #[test]
-    fn wal_wraps_after_checkpoint_and_recovers_across_the_lap() {
-        let path = unique_temp_path("wal-wrap");
-        let mut storage = create_small(&path);
-
-        // Fill ~60% of the 64 KiB ring, then reclaim it via checkpoint.
-        for seq in 1..=5u64 {
-            append_search_entry(&mut storage, seq, &vec![seq as u8; 8000]);
-        }
-        storage
-            .write_checkpoint(b"state-at-5", 5, 6, 1)
-            .expect("checkpoint");
-
-        // These cross the lap boundary (one entry forces a wrap gap).
-        let post: Vec<Vec<u8>> = (6..=9u64).map(|seq| vec![seq as u8; 8000]).collect();
-        for (i, payload) in post.iter().enumerate() {
-            append_search_entry(&mut storage, 6 + i as u64, payload);
-        }
-        drop(storage); // crash
-
-        let mut storage = Storage::open(path.clone()).expect("reopen");
-        let snapshot = storage
-            .load_snapshot()
-            .expect("load")
-            .expect("snapshot present");
-        assert_eq!(snapshot.data, b"state-at-5");
-        let records = storage.replay_wal_entries().expect("replay");
-        assert_eq!(
-            records.len(),
-            4,
-            "exactly the post-checkpoint entries replay"
-        );
-        for (record, payload) in records.iter().zip(&post) {
-            assert_eq!(&record.payload, payload);
-        }
-        // The ring stays usable after recovery on the wrapped lap.
-        append_search_entry(&mut storage, 10, b"after-recovery");
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-    }
-
-    #[test]
-    fn wal_full_errors_and_checkpoint_reclaims() {
-        let path = unique_temp_path("wal-full");
-        let mut storage = create_small(&path);
-        let payload = vec![7u8; 8000];
-        let mut appended = 0;
-        let err = loop {
-            match storage.append_wal_entry(WAL_ENTRY_TYPE_RECORD, 1, appended + 1, &payload) {
-                Ok(_) => appended += 1,
-                Err(err) => break err,
-            }
-            assert!(appended < 100, "ring must fill up");
-        };
-        assert!(matches!(err, StorageError::WalFull(_)), "got: {err}");
-
-        storage
-            .write_checkpoint(b"reclaim", appended, appended + 1, 1)
-            .expect("checkpoint");
-        append_search_entry(&mut storage, appended + 1, &payload);
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-    }
-
-    #[test]
-    fn extent_alloc_free_replays_from_wal() {
-        let path = unique_temp_path("extent-replay");
-        let mut storage = create_small(&path);
-        let durable = storage.allocate_extent(false).expect("durable extent");
-        let temp = storage.allocate_extent(true).expect("temp extent");
-        assert_ne!(durable, temp);
-        drop(storage); // crash: bitmap never persisted, only the WAL knows
-
-        let mut storage = Storage::open(path.clone()).expect("reopen");
-        assert!(
-            storage.is_page_allocated(durable),
-            "logged alloc must replay"
-        );
-        assert!(
-            !storage.is_page_allocated(temp),
-            "temp extents must vanish on restart"
-        );
-
-        storage.free_extent(durable).expect("free extent");
-        drop(storage); // crash again
-
-        let mut storage = Storage::open(path.clone()).expect("reopen after free");
-        assert!(
-            !storage.is_page_allocated(durable),
-            "logged free must replay"
-        );
-
-        // Alloc + checkpoint: the bitmap carries the state once the WAL is
-        // reclaimed.
-        let kept = storage.allocate_extent(false).expect("extent");
-        storage
-            .write_checkpoint(b"with-extent", 1, 2, 1)
-            .expect("checkpoint");
-        drop(storage);
-
-        let mut storage = Storage::open(path.clone()).expect("reopen after checkpoint");
-        assert!(
-            storage.replay_wal_entries().expect("replay").is_empty(),
-            "wal reclaimed"
-        );
-        assert!(
-            storage.is_page_allocated(kept),
-            "bitmap must carry extents across checkpoints"
-        );
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-    }
-
-    /// Regression (review finding): the whole-tail-page flush must never
-    /// let the page's zero suffix alias onto live entries at the head. With
-    /// a mid-page head and a nearly-full ring, the append that would have
-    /// clobbered the head must instead report WalFull, and every previously
-    /// acknowledged entry must survive a crash.
-    #[test]
-    fn tail_page_flush_never_overwrites_live_head_entries() {
-        let path = unique_temp_path("tail-page-alias");
-        let mut storage = create_small(&path);
-
-        // Head becomes mid-page: one 5056-byte entry, then checkpoint.
-        append_search_entry(&mut storage, 1, &vec![1u8; 5000]);
-        storage
-            .write_checkpoint(b"cp", 1, 2, 1)
-            .expect("checkpoint");
-
-        // Fill the ring almost entirely (the 15th entry wraps).
-        let mut acked = Vec::new();
-        for i in 0..15u64 {
-            let payload = vec![(i + 2) as u8; 4000];
-            append_search_entry(&mut storage, i + 2, &payload);
-            acked.push(payload);
-        }
-
-        // This append fits the naive byte count but its tail-page zero
-        // suffix would overwrite the oldest live entries; it must be
-        // rejected.
-        let err = storage
-            .append_wal_entry(WAL_ENTRY_TYPE_RECORD, 1, 17, &vec![9u8; 900])
-            .expect_err("append aliasing the head must fail");
-        assert!(matches!(err, StorageError::WalFull(_)), "got: {err}");
-        drop(storage); // crash
-
-        let mut storage = Storage::open(path.clone()).expect("reopen");
-        let records = storage.replay_wal_entries().expect("replay");
-        assert_eq!(records.len(), 15, "every acked entry must survive");
-        for (record, payload) in records.iter().zip(&acked) {
-            assert_eq!(&record.payload, payload);
-        }
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-    }
-
-    /// Regression (review finding): a crash between descriptor fsync and
-    /// superblock publish leaves descriptor generations ahead of the
-    /// superblocks. The next checkpoint must mint a strictly higher
-    /// generation (no duplicates), and later opens must agree on the live
-    /// snapshot.
-    #[test]
-    fn checkpoint_after_superblock_publish_crash_mints_higher_generation() {
-        let path = unique_temp_path("gen-minting");
-        let layout = test_layout();
-        let mut storage = create_small(&path);
-        storage.write_checkpoint(b"one", 1, 2, 1).expect("cp 1");
-        drop(storage);
-
-        // Save the checkpoint-1-era superblocks.
-        let sb_a = read_bytes(&path, layout.superblock_a_offset, SUPERBLOCK_SIZE);
-        let sb_b = read_bytes(&path, layout.superblock_b_offset, SUPERBLOCK_SIZE);
-
-        let mut storage = Storage::open(path.clone()).expect("reopen");
-        storage.write_checkpoint(b"two", 2, 3, 1).expect("cp 2");
-        drop(storage);
-
-        // Simulate the crash window: descriptor of checkpoint 2 durable,
-        // superblocks rolled back to checkpoint 1.
-        overwrite_bytes(&path, layout.superblock_a_offset, &sb_a);
-        overwrite_bytes(&path, layout.superblock_b_offset, &sb_b);
-
-        let mut storage = Storage::open(path.clone()).expect("reopen in crash window");
-        let snapshot = storage.load_snapshot().expect("load").expect("snapshot");
-        assert_eq!(snapshot.data, b"two", "newest descriptor must win");
-
-        // The next checkpoint must not duplicate checkpoint 2's generation.
-        storage.write_checkpoint(b"three", 3, 4, 1).expect("cp 3");
-        drop(storage);
-
-        let mut storage = Storage::open(path.clone()).expect("final reopen");
-        let snapshot = storage.load_snapshot().expect("load").expect("snapshot");
-        assert_eq!(snapshot.data, b"three");
-        // Allocator agrees with the snapshot choice: the live extent is
-        // allocated and loadable, and further checkpoints keep working.
-        storage.write_checkpoint(b"four", 4, 5, 1).expect("cp 4");
-        let snapshot = storage.load_snapshot().expect("load").expect("snapshot");
-        assert_eq!(snapshot.data, b"four");
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-    }
-
-    /// Review finding: a durable wrap gap whose next-lap entry never made it
-    /// to disk must rewind the recovered tail to the gap start.
-    #[test]
-    fn wrap_gap_without_next_lap_entry_rewinds_tail() {
-        let path = unique_temp_path("gap-rewind");
-        let layout = test_layout();
-        let mut storage = create_small(&path);
-        for seq in 1..=5u64 {
-            append_search_entry(&mut storage, seq, &vec![seq as u8; 8000]);
-        }
-        storage
-            .write_checkpoint(b"cp", 5, 6, 1)
-            .expect("checkpoint");
-        let post: Vec<Vec<u8>> = (6..=9u64).map(|seq| vec![seq as u8; 8000]).collect();
-        for (i, payload) in post.iter().enumerate() {
-            append_search_entry(&mut storage, 6 + i as u64, payload);
-        }
-        drop(storage);
-
-        // Entry 9 wrapped to the ring start. Erase its lap-2 pages as if the
-        // gap reached disk but the entry itself never did.
-        let entry_len = wal_entry_padded_len(8000);
-        overwrite_bytes(&path, layout.wal_offset, &vec![0u8; entry_len]);
-
-        let mut storage = Storage::open(path.clone()).expect("reopen");
-        let records = storage.replay_wal_entries().expect("replay");
-        assert_eq!(records.len(), 3, "the lost wrap entry must not replay");
-        for (record, payload) in records.iter().zip(&post[..3]) {
-            assert_eq!(&record.payload, payload);
-        }
-
-        // The rewound tail must be usable: a new append re-wraps and
-        // survives another crash.
-        append_search_entry(&mut storage, 9, b"after-rewind");
-        drop(storage);
-        let mut storage = Storage::open(path.clone()).expect("second reopen");
-        let records = storage.replay_wal_entries().expect("replay");
-        assert_eq!(records.len(), 4);
-        assert_eq!(records[3].payload, b"after-rewind");
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-    }
-
-    /// Review finding: the in-place lazy superblock rewrite stakes its
-    /// safety on falling back to the other superblock plus the forward
-    /// scan. Exercise exactly that: corrupt the active superblock and
-    /// recover everything through the stale one.
-    #[test]
-    fn torn_active_superblock_falls_back_and_forward_scan_recovers() {
-        let path = unique_temp_path("torn-superblock");
-        let layout = test_layout();
-        let mut storage = create_small(&path);
-        storage.lock().wal.set_superblock_interval(400);
-        let payloads: Vec<Vec<u8>> = (0..4u8).map(|i| vec![i + 1; 100]).collect();
-        for (i, payload) in payloads.iter().enumerate() {
-            append_search_entry(&mut storage, i as u64 + 1, payload);
-        }
-        drop(storage);
-
-        // The lazy writes all went to the active superblock (A). Tear it.
-        overwrite_bytes(
-            &path,
-            layout.superblock_a_offset,
-            &[0xEEu8; SUPERBLOCK_SIZE],
-        );
-
-        let mut storage = Storage::open(path.clone()).expect("reopen on backup superblock");
-        let records = storage.replay_wal_entries().expect("replay");
-        assert_eq!(records.len(), 4, "forward scan from the stale superblock");
-        for (record, payload) in records.iter().zip(&payloads) {
-            assert_eq!(&record.payload, payload);
-        }
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-    }
-
-    /// Corruption inside the trusted region truncates the log there; the
-    /// corrected tail must be persisted at open so stale entries beyond it
-    /// can never re-enter a future trusted region.
-    #[test]
-    fn trusted_region_corruption_truncates_and_persists_corrected_tail() {
-        let path = unique_temp_path("trusted-corruption");
-        let layout = test_layout();
-        let mut storage = create_small(&path);
-        storage.lock().wal.set_superblock_interval(400);
-        for seq in 1..=4u64 {
-            append_search_entry(&mut storage, seq, &[seq as u8; 100]);
-        }
-        drop(storage);
-        let entry_len = wal_entry_padded_len(100) as u64;
-
-        // Corrupt entry 2, inside the superblock-trusted region.
-        overwrite_bytes(&path, layout.wal_offset + entry_len + 8, &[0xDDu8; 32]);
-
-        let mut storage = Storage::open(path.clone()).expect("reopen after corruption");
-        let records = storage.replay_wal_entries().expect("replay");
-        assert_eq!(records.len(), 1, "log truncates at the corrupt entry");
-        assert_eq!(records[0].payload, vec![1u8; 100]);
-        drop(storage);
-
-        // The corrected (smaller) tail must now be on disk.
-        let sb = read_active_superblock(&path, &layout);
-        assert_eq!(
-            sb.wal_tail, entry_len,
-            "open must persist the corrected tail"
-        );
-
-        // New history: append a differently-sized entry and crash. Recovery
-        // must see [entry 1, new entry] and never resurrect old entries 3/4.
-        let mut storage = Storage::open(path.clone()).expect("reopen");
-        storage.replay_wal_entries().expect("drain");
-        append_search_entry(&mut storage, 2, &[9u8; 60]);
-        drop(storage);
-        let mut storage = Storage::open(path.clone()).expect("final reopen");
-        let records = storage.replay_wal_entries().expect("replay");
-        assert_eq!(records.len(), 2);
-        assert_eq!(records[0].payload, vec![1u8; 100]);
-        assert_eq!(records[1].payload, vec![9u8; 60]);
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-    }
-
-    /// Review finding: allocate_extent's rollback when the WAL append fails
-    /// must leave the allocator exactly as before.
-    #[test]
-    fn allocate_extent_rolls_back_when_wal_is_full() {
-        let path = unique_temp_path("extent-rollback");
-        let mut storage = create_small(&path);
-        // Fill the ring for large and small entries alike.
-        let mut seq = 1u64;
-        for payload_len in [8000usize, 1000, 100, 8] {
-            while storage
-                .append_wal_entry(WAL_ENTRY_TYPE_RECORD, 1, seq, &vec![7u8; payload_len])
-                .is_ok()
-            {
-                seq += 1;
-                assert!(seq < 1000, "ring must fill up");
-            }
-        }
-
-        let err = storage
-            .allocate_extent(false)
-            .expect_err("extent alloc must fail when its record cannot be logged");
-        assert!(matches!(err, StorageError::WalFull(_)), "got: {err}");
-        for page in 0..EXTENT_PAGES {
-            assert!(
-                !storage.is_page_allocated(page),
-                "rolled-back extent must leave page {page} free"
-            );
-        }
-
-        // After reclaiming the ring, extent allocation works again and the
-        // rolled-back range stays free (the next-fit cursor moved past it,
-        // so it is not the range reused here).
-        storage
-            .write_checkpoint(b"reclaim", seq, seq + 1, 1)
-            .expect("checkpoint");
-        let start = storage.allocate_extent(false).expect("extent");
-        for page in 0..EXTENT_PAGES {
-            assert!(!storage.is_page_allocated(page));
-        }
-        assert!(storage.is_page_allocated(start));
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-    }
-
-    #[test]
-    fn empty_checkpoint_data_is_rejected() {
-        let path = unique_temp_path("empty-checkpoint");
-        let mut storage = create_small(&path);
-        storage.write_checkpoint(b"valid", 1, 2, 1).expect("cp");
-        let err = storage
-            .write_checkpoint(b"", 2, 3, 1)
-            .expect_err("empty checkpoint must be rejected");
-        assert!(matches!(err, StorageError::InvalidConfig(_)), "got: {err}");
-        // The previous snapshot is untouched.
-        let snapshot = storage.load_snapshot().expect("load").expect("snapshot");
-        assert_eq!(snapshot.data, b"valid");
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-    }
-
-    /// Review finding: the v1 upgrade must refresh BOTH superblocks — a v1
-    /// backup superblock is arbitrarily stale, and a later fallback to it
-    /// would lose every v1 WAL entry (they cannot pass the v2 forward scan).
-    #[test]
-    fn upgraded_v1_file_survives_active_superblock_loss() {
-        let path = unique_temp_path("v1-superblock-refresh");
-        let events = vec![(1u64, vec![5u8; 80]), (2u64, vec![6u8; 40])];
-        let layout = write_v1_fixture(&path, &events, None);
-
-        // First open performs the upgrade.
-        let mut storage = Storage::open(path.clone()).expect("upgrade open");
-        assert_eq!(storage.replay_wal_entries().expect("replay").len(), 2);
-        drop(storage);
-
-        // Lose the active superblock; the refreshed backup must carry the
-        // v1 tail so the trusted scan still finds the v1-stamped entries.
-        overwrite_bytes(
-            &path,
-            layout.superblock_a_offset,
-            &[0xEEu8; SUPERBLOCK_SIZE],
-        );
-        let mut storage = Storage::open(path.clone()).expect("reopen on backup");
-        let records = storage.replay_wal_entries().expect("replay");
-        assert_eq!(records.len(), 2, "v1 entries must survive the fallback");
-        assert_eq!(records[0].payload, vec![5u8; 80]);
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-    }
-
-    #[test]
-    fn successive_checkpoints_recycle_snapshot_extents() {
-        let path = unique_temp_path("snapshot-recycle");
-        let mut storage = create_small(&path);
-        storage.write_checkpoint(b"first", 1, 2, 1).expect("cp 1");
-        let first_desc = storage
-            .lock()
-            .load_active_snapshot_descriptor()
-            .expect("desc")
-            .expect("present");
-        let (first_start, first_pages) = storage
-            .lock()
-            .descriptor_page_range(&first_desc)
-            .expect("range");
-        assert!(storage.is_page_allocated(first_start));
-
-        storage
-            .write_checkpoint(b"second-snapshot", 2, 3, 1)
-            .expect("cp 2");
-        for page in first_start..first_start + first_pages {
-            assert!(
-                !storage.is_page_allocated(page),
-                "first snapshot extent must be freed after the second checkpoint"
-            );
-        }
-        drop(storage);
-
-        let mut storage = Storage::open(path.clone()).expect("reopen");
-        let snapshot = storage.load_snapshot().expect("load").expect("snapshot");
-        assert_eq!(snapshot.data, b"second-snapshot");
-        assert!(!storage.is_page_allocated(first_start));
-        drop(storage);
-        let _ = std::fs::remove_file(path);
-    }
-}
+mod tests;
